% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_10.R
\name{add_image_slide}
\alias{add_image_slide}
\title{Perform add image slide operation}
\usage{
add_image_slide(ppt, image1, image2, h1 = "heading 1", h2 = "heading 2")
}
\arguments{
\item{ppt}{Parameter for ppt}

\item{image1}{Parameter for image1}

\item{image2}{Parameter for image2}

\item{h1}{Character string (default: "heading 1")}

\item{h2}{Character string (default: "heading 2")}
}
\value{
Function result
}
\description{
Perform add image slide operation
}
