% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_3.R
\name{add_phase_to_filtered_seu}
\alias{add_phase_to_filtered_seu}
\title{Add cell cycle Phase to a filtered Seurat object and save}
\usage{
add_phase_to_filtered_seu(seu_path, organism = "human")
}
\arguments{
\item{seu_path}{File path to filtered Seurat RDS}

\item{organism}{Character string (default: "human")}
}
\value{
Path to updated RDS file
}
\description{
Add cell cycle Phase to a filtered Seurat object and save
}
