% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_8.R
\name{all_same_sign}
\alias{all_same_sign}
\title{Perform all same sign operation}
\usage{
all_same_sign(x)
}
\arguments{
\item{x}{Parameter for x}
}
\value{
Function result
}
\description{
Perform all same sign operation
}
