% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_10.R
\name{annotables_to_grange}
\alias{annotables_to_grange}
\title{Perform annotables to grange operation}
\usage{
annotables_to_grange()
}
\value{
Function result
}
\description{
Perform annotables to grange operation
}
