% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_20.R
\name{annotate_cell_cycle_singleR}
\alias{annotate_cell_cycle_singleR}
\title{Add annotations to data}
\usage{
annotate_cell_cycle_singleR(seu)
}
\arguments{
\item{seu}{Seurat object}
}
\value{
Modified Seurat object
}
\description{
Add annotations to data
}
