% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_12.R
\name{annotate_cell_cycle_without_1q}
\alias{annotate_cell_cycle_without_1q}
\title{Add annotations to data}
\usage{
annotate_cell_cycle_without_1q(seu, organism = "human", ...)
}
\arguments{
\item{seu}{Seurat object}

\item{organism}{Character string (default: "human")}

\item{...}{Additional arguments passed to other functions}
}
\value{
Function result
}
\description{
Add annotations to data
}
