% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_10.R
\name{annotate_cluster_membership}
\alias{annotate_cluster_membership}
\title{Perform clustering analysis}
\usage{
annotate_cluster_membership(diffex_1, diffex_2, new_col_name)
}
\arguments{
\item{diffex_1}{Parameter for diffex 1}

\item{diffex_2}{Parameter for diffex 2}

\item{new_col_name}{Color specification}
}
\value{
Function result
}
\description{
Perform clustering analysis
}
