% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_12.R
\name{annotate_cluster_scna_percentage}
\alias{annotate_cluster_scna_percentage}
\title{Perform clustering analysis}
\usage{
annotate_cluster_scna_percentage(seu)
}
\arguments{
\item{seu}{Seurat object}
}
\value{
Modified Seurat object
}
\description{
Perform clustering analysis
}
