% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_3.R
\name{annotate_filter_reason}
\alias{annotate_filter_reason}
\title{Filter data based on specified criteria}
\usage{
annotate_filter_reason(
  seu,
  sample_id,
  cluster_dictionary,
  cells_to_remove = NULL,
  mito_threshold = 10,
  nCount_threshold = 1000,
  nFeature_threshold = 1000
)
}
\arguments{
\item{cluster_dictionary}{Cluster information}

\item{cells_to_remove}{Cell identifiers or information}

\item{numbat_rds_file}{File path}

\item{seus}{Parameter for seus}

\item{large_clone_simplifications}{Parameter for large clone simplifications}

\item{filter_expressions}{Parameter for filter expressions}

\item{extension}{Character string (default: "")}

\item{leiden_cluster_file}{File path}
}
\value{
Modified Seurat object
}
\description{
Filter data based on specified criteria
}
