% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_21.R
\name{annotate_seu_with_rb_subtype_gene_expression}
\alias{annotate_seu_with_rb_subtype_gene_expression}
\title{Add annotations to data}
\usage{
annotate_seu_with_rb_subtype_gene_expression(seu)
}
\arguments{
\item{seu}{Seurat object}

\item{hallmark_gene_set}{Gene names or identifiers}
}
\value{
Modified Seurat object

Extracted data elements
}
\description{
Add annotations to data
}
