% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_38.R
\name{append_clone_nums}
\alias{append_clone_nums}
\title{Perform append clone nums operation}
\usage{
append_clone_nums(diffex, clone_comparison, seu)
}
\arguments{
\item{diffex}{Parameter for diffex}

\item{clone_comparison}{Parameter for clone comparison}

\item{seu}{Seurat object}
}
\value{
Function result
}
\description{
Perform append clone nums operation
}
