% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_8.R
\name{assign_designated_phase_clusters}
\alias{assign_designated_phase_clusters}
\title{Perform clustering analysis}
\usage{
assign_designated_phase_clusters(
  seu_path,
  cluster_orders,
  resolution_dictionary
)
}
\arguments{
\item{seu_path}{File path}

\item{cluster_orders}{Cluster information}

\item{resolution_dictionary}{Parameter for resolution dictionary}
}
\value{
Function result
}
\description{
Perform clustering analysis
}
