% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_50.R
\name{assign_phase_clusters}
\alias{assign_phase_clusters}
\title{Perform clustering analysis}
\usage{
assign_phase_clusters(
  seu = NULL,
  file_id = NULL,
  cluster_orders = NULL,
  resolution = "0",
  group.by = "SCT_snn_res.0.6"
)
}
\arguments{
\item{seu}{Seurat object}

\item{file_id}{File path}

\item{cluster_orders}{Cluster information}

\item{resolution}{Character string (default: "0")}

\item{group.by}{Character string (default: "SCT_snn_res.0.6")}
}
\value{
Modified Seurat object
}
\description{
Perform clustering analysis
}
