% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_14.R
\name{binom_test_clone_percent}
\alias{binom_test_clone_percent}
\title{Perform binom test clone percent operation}
\usage{
binom_test_clone_percent(mydf)
}
\arguments{
\item{mydf}{Parameter for mydf}
}
\value{
Function result
}
\description{
Perform binom test clone percent operation
}
