% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_29.R
\name{browse_celltype_expression}
\alias{browse_celltype_expression}
\title{Perform browse celltype expression operation}
\usage{
browse_celltype_expression(sridhar_seu, symbol)
}
\arguments{
\item{sridhar_seu}{Parameter for sridhar seu}

\item{symbol}{Parameter for symbol}
}
\value{
Data frame
}
\description{
Perform browse celltype expression operation
}
