% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_52.R
\name{calc_silhouette}
\alias{calc_silhouette}
\title{Perform calc silhouette operation}
\usage{
calc_silhouette(seu_path, assay = "SCT", reduction = "pca", dims = 1:30, ...)
}
\arguments{
\item{seu_path}{File path}

\item{assay}{Character string (default: "SCT")}

\item{reduction}{Character string (default: "pca")}

\item{dims}{Parameter for dims}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Perform calc silhouette operation
}
