% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_14.R
\name{calculate_6p_percent_affected}
\alias{calculate_6p_percent_affected}
\title{Calculate or compute values}
\usage{
calculate_6p_percent_affected()
}
\value{
Calculated values
}
\description{
Calculate or compute values
}
