% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_14.R
\name{calculate_arm_percent}
\alias{calculate_arm_percent}
\title{Calculate or compute values}
\usage{
calculate_arm_percent(df)
}
\arguments{
\item{df}{Input data frame or dataset}
}
\value{
Calculated values
}
\description{
Calculate or compute values
}
