% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_10.R
\name{calculate_avg_cluster_distance}
\alias{calculate_avg_cluster_distance}
\title{Perform clustering analysis}
\usage{
calculate_avg_cluster_distance(x = "0", y = "1", df_list)
}
\arguments{
\item{x}{Character string (default: "0")}

\item{y}{Character string (default: "1")}

\item{df_list}{Parameter for df list}
}
\value{
Calculated values
}
\description{
Perform clustering analysis
}
