% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_13.R
\name{check_cluster_marker_gene}
\alias{check_cluster_marker_gene}
\title{Perform clustering analysis}
\usage{
check_cluster_marker_gene(seu_path, cluster_dictionary)
}
\arguments{
\item{seu_path}{File path}

\item{cluster_dictionary}{Cluster information}
}
\value{
Data frame
}
\description{
Perform clustering analysis
}
