% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_8.R
\name{check_merged_metadata}
\alias{check_merged_metadata}
\title{Merge or join datasets}
\usage{
check_merged_metadata(merged_seu_path, child_seu_paths)
}
\arguments{
\item{merged_seu_path}{File path}

\item{child_seu_paths}{File path}
}
\value{
Modified Seurat object
}
\description{
Merge or join datasets
}
