% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_20.R
\name{check_MYCN_mean_expression}
\alias{check_MYCN_mean_expression}
\title{Perform check mycn mean expression operation}
\usage{
check_MYCN_mean_expression(seu_path)
}
\arguments{
\item{seu_path}{File path}
}
\value{
Function result
}
\description{
Perform check mycn mean expression operation
}
