% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_46.R
\name{chi_sq_daughter_clusters}
\alias{chi_sq_daughter_clusters}
\title{Perform clustering analysis}
\usage{
chi_sq_daughter_clusters(
  seu,
  daughter_clusters,
  resolution,
  from_clust,
  assay = "SCT"
)
}
\arguments{
\item{seu}{Seurat object}

\item{daughter_clusters}{Cluster information}

\item{resolution}{Parameter for resolution}

\item{from_clust}{Parameter for from clust}

\item{assay}{Character string (default: "SCT")}
}
\value{
Function result
}
\description{
Perform clustering analysis
}
