% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_51.R
\name{clean_diffex}
\alias{clean_diffex}
\title{Perform differential expression analysis}
\usage{
clean_diffex(diffex, p_val_limit = 0.05, log2fc_limit = 1)
}
\arguments{
\item{diffex}{Parameter for diffex}

\item{p_val_limit}{Parameter for p val limit}

\item{log2fc_limit}{Parameter for log2fc limit}
}
\value{
Differential expression results
}
\description{
Perform differential expression analysis
}
