% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_10.R
\name{clone_cc_plots_by_scna}
\alias{clone_cc_plots_by_scna}
\title{Create a plot visualization}
\usage{
clone_cc_plots_by_scna(seu_list, scna_of_interest = "1q", ...)
}
\arguments{
\item{seu_list}{Parameter for seu list}

\item{scna_of_interest}{Character string (default: "1q")}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
