% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_25.R
\name{clone_diff_per_cluster}
\alias{clone_diff_per_cluster}
\title{Perform clustering analysis}
\usage{
clone_diff_per_cluster(cluster_for_diffex, seu, group.by)
}
\arguments{
\item{cluster_for_diffex}{Cluster information}

\item{seu}{Seurat object}

\item{group.by}{Parameter for group.by}
}
\value{
Function result
}
\description{
Perform clustering analysis
}
