% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_47.R
\name{collate_clone_distribution_tables}
\alias{collate_clone_distribution_tables}
\title{Perform collate clone distribution tables operation}
\usage{
collate_clone_distribution_tables(
  clone_distribution_plots,
  interesting_samples
)
}
\arguments{
\item{clone_distribution_plots}{Parameter for clone distribution plots}

\item{interesting_samples}{Parameter for interesting samples}
}
\value{
List object
}
\description{
Perform collate clone distribution tables operation
}
