% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_29.R
\name{collect_all_markers}
\alias{collect_all_markers}
\title{Perform collect all markers operation}
\usage{
collect_all_markers(numbat_rds_files, excel_output = "results/markers.xlsx")
}
\arguments{
\item{numbat_rds_files}{File path}

\item{excel_output}{Character string (default: "results/markers.xlsx")}
}
\value{
Function result
}
\description{
Perform collect all markers operation
}
