% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_29.R
\name{collect_markers}
\alias{collect_markers}
\title{Perform collect markers operation}
\usage{
collect_markers(
  numbat_rds_file,
  metavar = "gene_snn_res.0.2",
  num_markers = 5,
  selected_values = NULL,
  return_plotly = TRUE,
  marker_method = "presto",
  seurat_assay = "gene",
  hide_technical = "all",
  unique_markers = FALSE,
  p_val_cutoff = 1,
  ...
)
}
\arguments{
\item{numbat_rds_file}{File path}

\item{metavar}{Character string (default: "gene_snn_res.0.2")}

\item{num_markers}{Parameter for num markers}

\item{selected_values}{Parameter for selected values}

\item{return_plotly}{Logical flag (default: TRUE)}

\item{marker_method}{Character string (default: "presto")}

\item{seurat_assay}{Character string (default: "gene")}

\item{hide_technical}{Character string (default: "all")}

\item{unique_markers}{Logical flag (default: FALSE)}

\item{p_val_cutoff}{Threshold value for filtering}

\item{...}{Additional arguments passed to other functions}
}
\value{
Function result
}
\description{
Perform collect markers operation
}
