% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_10.R
\name{collect_study_metadata}
\alias{collect_study_metadata}
\title{Perform collect study metadata operation}
\usage{
collect_study_metadata()
}
\value{
Function result
}
\description{
Perform collect study metadata operation
}
