% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_45.R
\name{color_clustree_by_clone}
\alias{color_clustree_by_clone}
\title{Perform color clustree by clone operation}
\usage{
color_clustree_by_clone(
  clustree_plot,
  mycolor,
  myclone,
  mylabel = "asdf",
  sample_id
)
}
\arguments{
\item{clustree_plot}{Parameter for clustree plot}

\item{mycolor}{Color specification}

\item{myclone}{Parameter for myclone}

\item{mylabel}{Character string (default: "asdf")}

\item{sample_id}{Parameter for sample id}
}
\value{
ggplot2 plot object
}
\description{
Perform color clustree by clone operation
}
