% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_11.R
\name{compare_corresponding_enrichments}
\alias{compare_corresponding_enrichments}
\title{Perform enrichment analysis}
\usage{
compare_corresponding_enrichments(
  enrichment_list,
  common_seus = NULL,
  plot_path = tempfile(tmpdir = "results", fileext = ".pdf"),
  ...
)
}
\arguments{
\item{enrichment_list}{Parameter for enrichment list}

\item{common_seus}{Parameter for common seus}

\item{plot_path}{File path}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Perform enrichment analysis
}
