% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions_2.R
\name{compare_infercnv}
\alias{compare_infercnv}
\title{Perform compare infercnv operation}
\usage{
compare_infercnv(myseus, sample_id)
}
\arguments{
\item{myseus}{Parameter for myseus}

\item{sample_id}{Parameter for sample id}
}
\value{
Function result
}
\description{
Perform compare infercnv operation
}
