% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_2.R
\name{compare_per_cluster_and_total_clone_diffex}
\alias{compare_per_cluster_and_total_clone_diffex}
\title{Perform differential expression analysis}
\usage{
compare_per_cluster_and_total_clone_diffex(
  per_cluster_clone_diffex,
  total_clone_diffex
)
}
\arguments{
\item{per_cluster_clone_diffex}{Cluster information}

\item{total_clone_diffex}{Parameter for total clone diffex}
}
\value{
Differential expression results
}
\description{
Perform differential expression analysis
}
