% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_36.R
\name{compile_cis_trans_enrichment_recurrence_by_cluster}
\alias{compile_cis_trans_enrichment_recurrence_by_cluster}
\title{Perform enrichment analysis}
\usage{
compile_cis_trans_enrichment_recurrence_by_cluster(
  enrichment_tables,
  cis_plot_file = "results/cis_enrichment_plots",
  trans_plot_file = "results/trans_enrichment_plots",
  cis_table_file = "results/cis_enrichment_tables.xlsx",
  trans_table_file = "results/trans_enrichment_tables.xlsx",
  ...
)
}
\arguments{
\item{enrichment_tables}{Parameter for enrichment tables}

\item{cis_plot_file}{File path}

\item{trans_plot_file}{File path}

\item{cis_table_file}{File path}

\item{trans_table_file}{File path}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Perform enrichment analysis
}
