% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_7.R
\name{compplot_feature_and_clusters}
\alias{compplot_feature_and_clusters}
\title{Generate a feature plot}
\usage{
compplot_feature_and_clusters(seu, feature)
}
\arguments{
\item{seu}{Seurat object}

\item{feature}{Parameter for feature}
}
\value{
ggplot2 plot object
}
\description{
Generate a feature plot
}
