% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions_2.R
\name{convert_numbat_pngs}
\alias{convert_numbat_pngs}
\title{Perform numbat-related analysis}
\usage{
convert_numbat_pngs(numbat_rds_file, n_sample = 2000, lim = 0.8)
}
\arguments{
\item{numbat_rds_file}{File path}
}
\value{
Data frame
}
\description{
Perform numbat-related analysis
}
