% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_42.R
\name{convert_seu_to_scanpy}
\alias{convert_seu_to_scanpy}
\title{Convert data from one format to another}
\usage{
convert_seu_to_scanpy(seu_path)
}
\arguments{
\item{seu_path}{File path}
}
\value{
Converted data object
}
\description{
Convert data from one format to another
}
