% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_12.R
\name{convert_v5_to_v3}
\alias{convert_v5_to_v3}
\title{Convert data from one format to another}
\usage{
convert_v5_to_v3(seu_v5)
}
\arguments{
\item{seu_v5}{Parameter for seu v5}
}
\value{
Modified Seurat object
}
\description{
Convert data from one format to another
}
