% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_33.R
\name{convert_volcano_to_plotly}
\alias{convert_volcano_to_plotly}
\title{Generate a volcano plot for differential expression data}
\usage{
convert_volcano_to_plotly(myplot, out_html)
}
\arguments{
\item{myplot}{Plot object (ggplot2)}

\item{out_html}{Parameter for out html}
}
\value{
ggplot2 plot object
}
\description{
Generate a volcano plot for differential expression data
}
