% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_1.R
\name{debranch_seus}
\alias{debranch_seus}
\title{Perform debranch seus operation}
\usage{
debranch_seus(filtered_seus, branch_dictionary, ...)
}
\arguments{
\item{filtered_seus}{Parameter for filtered seus}

\item{branch_dictionary}{Parameter for branch dictionary}

\item{...}{Additional arguments passed to other functions}
}
\value{
Modified Seurat object
}
\description{
Perform debranch seus operation
}
