% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_39.R
\name{derive_pseudobulk_subtype_scores}
\alias{derive_pseudobulk_subtype_scores}
\title{Calculate scores for the given data}
\usage{
derive_pseudobulk_subtype_scores(seu_paths)
}
\arguments{
\item{seu_paths}{File path}

\item{seu_path}{File path}
}
\value{
ggplot2 plot object

Data frame
}
\description{
Calculate scores for the given data
}
