% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions_1.R
\name{diffex_by_cluster}
\alias{diffex_by_cluster}
\title{Perform differential expression analysis}
\usage{
diffex_by_cluster(sample_id, myseus, celldf, ...)
}
\arguments{
\item{sample_id}{Parameter for sample id}

\item{myseus}{Parameter for myseus}

\item{celldf}{Cell identifiers or information}

\item{...}{Additional arguments passed to other functions}

\item{cluster}{Cluster information}

\item{seu}{Seurat object}
}
\value{
Differential expression results

Filtered data
}
\description{
Perform differential expression analysis
}
