% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_1.R
\name{diffex_cells}
\alias{diffex_cells}
\title{Perform differential expression analysis}
\usage{
diffex_cells(myseus, cells.1, cells.2, ...)
}
\arguments{
\item{myseus}{Parameter for myseus}

\item{cells.1}{Cell identifiers or information}

\item{cells.2}{Cell identifiers or information}

\item{...}{Additional arguments passed to other functions}
}
\value{
Differential expression results
}
\description{
Perform differential expression analysis
}
