% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_1.R
\name{diffex_groups}
\alias{diffex_groups}
\title{Perform differential expression analysis}
\usage{
diffex_groups(numbat_rds_file, filter_expressions, idents = NULL)
}
\arguments{
\item{numbat_rds_file}{File path}

\item{filter_expressions}{Parameter for filter expressions}

\item{idents}{Cell identities or groups}
}
\value{
ggplot2 plot object
}
\description{
Perform differential expression analysis
}
