% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_23.R
\name{diffex_groups_old}
\alias{diffex_groups_old}
\title{Perform differential expression analysis}
\usage{
diffex_groups_old(sample_id, myseus, celldf, ...)
}
\arguments{
\item{sample_id}{Parameter for sample id}

\item{myseus}{Parameter for myseus}

\item{celldf}{Cell identifiers or information}

\item{...}{Additional arguments passed to other functions}
}
\value{
Differential expression results
}
\description{
Perform differential expression analysis
}
