% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_18.R
\name{diffex_per_cluster}
\alias{diffex_per_cluster}
\title{Perform differential expression analysis}
\usage{
diffex_per_cluster(
  seu,
  ident_1 = "w_scna",
  ident_2 = "wo_scna",
  mycluster = "g1_1",
  segments = NULL
)
}
\arguments{
\item{seu}{Seurat object}

\item{ident_1}{Cell identities or groups}

\item{ident_2}{Cell identities or groups}

\item{mycluster}{Cluster information}

\item{segments}{Parameter for segments}
}
\value{
Differential expression results
}
\description{
Perform differential expression analysis
}
