% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions_1.R
\name{drop_cluster_orders_table}
\alias{drop_cluster_orders_table}
\title{Drop and recreate cluster_orders table
Use this if you need to clear old non-JSON data}
\usage{
drop_cluster_orders_table(sqlite_path = "batch_hashes.sqlite")
}
\description{
Drop and recreate cluster_orders table
Use this if you need to clear old non-JSON data
}
