% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_14.R
\name{drop_mt_cluster}
\alias{drop_mt_cluster}
\title{Perform clustering analysis}
\usage{
drop_mt_cluster(seu, group.by = "integrated_snn_res.0.4")
}
\arguments{
\item{seu}{Seurat object}

\item{group.by}{Character string (default: "integrated_snn_res.0.4")}
}
\value{
Modified Seurat object
}
\description{
Perform clustering analysis
}
