% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_12.R
\name{drop_y_axis}
\alias{drop_y_axis}
\title{Perform drop y axis operation}
\usage{
drop_y_axis(myplot)
}
\arguments{
\item{myplot}{Plot object (ggplot2)}
}
\value{
ggplot2 plot object
}
\description{
Perform drop y axis operation
}
