% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions_2.R
\name{enrich_by_cluster}
\alias{enrich_by_cluster}
\title{Perform enrichment analysis}
\usage{
enrich_by_cluster(seu_path)
}
\arguments{
\item{seu_path}{File path}

\item{df}{Input data frame or dataset}

\item{cc.genes}{Gene names or identifiers}
}
\value{
ggplot2 plot object

Function result
}
\description{
Perform enrichment analysis
}
