% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions_2.R
\name{enrich_diffex_by_cluster}
\alias{enrich_diffex_by_cluster}
\title{Perform differential expression analysis}
\usage{
enrich_diffex_by_cluster(sample_id, myseus, celldf, ...)
}
\arguments{
\item{sample_id}{Parameter for sample id}

\item{myseus}{Parameter for myseus}

\item{celldf}{Cell identifiers or information}

\item{...}{Additional arguments passed to other functions}
}
\value{
Data frame
}
\description{
Perform differential expression analysis
}
