% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_35.R
\name{enrich_oncoprints_clusters}
\alias{enrich_oncoprints_clusters}
\title{Perform enrichment analysis}
\usage{
enrich_oncoprints_clusters(
  large_filter_expressions,
  cluster_dictionary,
  interesting_samples,
  cis_diffex_clones_for_each_cluster,
  trans_diffex_clones_for_each_cluster,
  large_clone_comparisons,
  ...
)
}
\arguments{
\item{large_filter_expressions}{Parameter for large filter expressions}

\item{cluster_dictionary}{Cluster information}

\item{interesting_samples}{Parameter for interesting samples}

\item{cis_diffex_clones_for_each_cluster}{Cluster information}

\item{trans_diffex_clones_for_each_cluster}{Cluster information}

\item{large_clone_comparisons}{Parameter for large clone comparisons}

\item{...}{Additional arguments passed to other functions}
}
\value{
List object
}
\description{
Perform enrichment analysis
}
