% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_19.R
\name{extract_full_segmentation}
\alias{extract_full_segmentation}
\title{Extract or pull specific data elements}
\usage{
extract_full_segmentation(
  numbat_rds_files,
  table_path = "results/table_s11.xlsx"
)
}
\arguments{
\item{numbat_rds_files}{File path}

\item{table_path}{File path}
}
\value{
Extracted data elements
}
\description{
Extract or pull specific data elements
}
