% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_3.R
\name{filter_input_by_region}
\alias{filter_input_by_region}
\title{Filter data based on specified criteria}
\usage{
filter_input_by_region(df, segment_region, oncoprint_settings)
}
\arguments{
\item{df}{Input data frame or dataset}

\item{segment_region}{Parameter for segment region}
}
\value{
Filtered data
}
\description{
Filter data based on specified criteria
}
