% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_33.R
\name{filter_input_by_scna}
\alias{filter_input_by_scna}
\title{Filter data based on specified criteria}
\usage{
filter_input_by_scna(
  df,
  scna_of_interest,
  p_val_threshold,
  fc_threshold,
  n_recur
)
}
\arguments{
\item{df}{Input data frame or dataset}

\item{scna_of_interest}{Parameter for scna of interest}

\item{p_val_threshold}{Threshold value for filtering}

\item{fc_threshold}{Threshold value for filtering}

\item{n_recur}{Parameter for n recur}
}
\value{
Filtered data
}
\description{
Filter data based on specified criteria
}
