% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_7.R
\name{filter_numbat_cells}
\alias{filter_numbat_cells}
\title{Filter data based on specified criteria}
\usage{
filter_numbat_cells(sample_id, myseus, mynbs, merged_metadata, myexpressions)
}
\arguments{
\item{sample_id}{Parameter for sample id}

\item{myseus}{Parameter for myseus}

\item{mynbs}{Parameter for mynbs}

\item{merged_metadata}{Parameter for merged metadata}

\item{myexpressions}{Parameter for myexpressions}
}
\value{
Filtered data
}
\description{
Filter data based on specified criteria
}
