% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_3.R
\name{filter_oncoprint_diffex}
\alias{filter_oncoprint_diffex}
\title{Filter data based on specified criteria}
\usage{
filter_oncoprint_diffex(unfiltered_oncoprint_input_by_scna, oncoprint_settings)
}
\arguments{
\item{unfiltered_oncoprint_input_by_scna}{Parameter for unfiltered oncoprint input by scna}

\item{oncoprint_settings}{Parameter for oncoprint settings}
}
\value{
Filtered data
}
\description{
Filter data based on specified criteria
}
