% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_5.R
\name{filter_phylo_plot}
\alias{filter_phylo_plot}
\title{Create a plot visualization}
\usage{
filter_phylo_plot(nb, myseu, myannot, mytitle, expressions, ...)
}
\arguments{
\item{nb}{Numbat object}

\item{myseu}{Seurat object}

\item{myannot}{Parameter for myannot}

\item{mytitle}{Plot title}

\item{expressions}{Parameter for expressions}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
