% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_48.R
\name{find_all_diffex_from_clustree}
\alias{find_all_diffex_from_clustree}
\title{Perform differential expression analysis}
\usage{
find_all_diffex_from_clustree(table_set, debranched_seus, ...)
}
\arguments{
\item{table_set}{Parameter for table set}

\item{debranched_seus}{Parameter for debranched seus}

\item{...}{Additional arguments passed to other functions}

\item{to_clust}{Parameter for to clust}

\item{to_SCT_snn_res.}{Parameter for to SCT snn res.}

\item{seu}{Seurat object}
}
\value{
Differential expression results

Function result
}
\description{
Perform differential expression analysis
}
