% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_49.R
\name{find_candidate_cis_in_clustree_diffexes}
\alias{find_candidate_cis_in_clustree_diffexes}
\title{Perform differential expression analysis}
\usage{
find_candidate_cis_in_clustree_diffexes(clustree_diffexes)
}
\arguments{
\item{clustree_diffexes}{Parameter for clustree diffexes}
}
\value{
ggplot2 plot object
}
\description{
Perform differential expression analysis
}
