% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_49.R
\name{find_candidate_trans_in_clustree_diffexes}
\alias{find_candidate_trans_in_clustree_diffexes}
\title{Perform differential expression analysis}
\usage{
find_candidate_trans_in_clustree_diffexes(
  clustree_diffexes,
  gene_location = "trans"
)
}
\arguments{
\item{clustree_diffexes}{Parameter for clustree diffexes}

\item{gene_location}{Gene names or identifiers}
}
\value{
ggplot2 plot object
}
\description{
Perform differential expression analysis
}
