% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_52.R
\name{find_cc_genes_by_arm}
\alias{find_cc_genes_by_arm}
\title{Perform find cc genes by arm operation}
\usage{
find_cc_genes_by_arm()
}
\value{
Function result
}
\description{
Perform find cc genes by arm operation
}
