% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_26.R
\name{find_diffex_bw_clones_for_each_cluster}
\alias{find_diffex_bw_clones_for_each_cluster}
\title{Perform differential expression analysis}
\usage{
find_diffex_bw_clones_for_each_cluster(
  seu_path,
  numbat_rds_files,
  large_clone_comparisons,
  cluster_dictionary,
  location = "cis",
  scna_of_interest = NULL
)
}
\arguments{
\item{seu_path}{File path}

\item{numbat_rds_files}{File path}

\item{large_clone_comparisons}{Parameter for large clone comparisons}

\item{cluster_dictionary}{Cluster information}

\item{location}{Character string (default: "cis")}

\item{scna_of_interest}{Parameter for scna of interest}
}
\value{
Differential expression results
}
\description{
Perform differential expression analysis
}
