% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_25.R
\name{find_diffex_bw_clusters_for_each_clone}
\alias{find_diffex_bw_clusters_for_each_clone}
\title{Perform differential expression analysis}
\usage{
find_diffex_bw_clusters_for_each_clone(
  numbat_rds_file,
  cluster_dictionary,
  ident.1 = "G2M",
  ident.2 = "cone"
)
}
\arguments{
\item{numbat_rds_file}{File path}

\item{cluster_dictionary}{Cluster information}

\item{ident.1}{Cell identities or groups}

\item{ident.2}{Cell identities or groups}

\item{clone_for_diffex}{Parameter for clone for diffex}

\item{seu}{Seurat object}

\item{diffex_bw_clusters_for_each_clone}{Cluster information}
}
\value{
ggplot2 plot object

Function result

Differential expression results
}
\description{
Perform differential expression analysis
}
