% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_30.R
\name{find_diffex_clones}
\alias{find_diffex_clones}
\title{Perform differential expression analysis}
\usage{
find_diffex_clones(
  seu_path,
  numbat_rds_files,
  large_clone_comparisons,
  location = "cis"
)
}
\arguments{
\item{seu_path}{File path}

\item{numbat_rds_files}{File path}

\item{large_clone_comparisons}{Parameter for large clone comparisons}

\item{location}{Character string (default: "cis")}
}
\value{
Differential expression results
}
\description{
Perform differential expression analysis
}
