% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_30.R
\name{find_diffex_clones_in_phase}
\alias{find_diffex_clones_in_phase}
\title{Perform differential expression analysis}
\usage{
find_diffex_clones_in_phase(
  seu_path,
  phase = "g1",
  scna_of_interest = "2p",
  numbat_rds_files,
  large_clone_comparisons,
  location = "cis"
)
}
\arguments{
\item{seu_path}{File path}

\item{phase}{Character string (default: "g1")}

\item{scna_of_interest}{Character string (default: "2p")}

\item{numbat_rds_files}{File path}

\item{large_clone_comparisons}{Parameter for large clone comparisons}

\item{location}{Character string (default: "cis")}
}
\value{
Differential expression results
}
\description{
Perform differential expression analysis
}
