% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_31.R
\name{find_diffex_clones_integrated}
\alias{find_diffex_clones_integrated}
\title{Perform differential expression analysis}
\usage{
find_diffex_clones_integrated(
  seu_path,
  kept_samples,
  clone_comparisons,
  location = "cis"
)
}
\arguments{
\item{seu_path}{File path}

\item{kept_samples}{Parameter for kept samples}

\item{clone_comparisons}{Parameter for clone comparisons}

\item{location}{Character string (default: "cis")}
}
\value{
Differential expression results
}
\description{
Perform differential expression analysis
}
