% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_31.R
\name{find_diffex_clusters_between_corresponding_states}
\alias{find_diffex_clusters_between_corresponding_states}
\title{Perform differential expression analysis}
\usage{
find_diffex_clusters_between_corresponding_states(
  seu_path,
  corresponding_states_dictionary,
  large_clone_comparisons = large_clone_comparisons,
  numbat_rds_files = numbat_rds_files,
  location = "cis"
)
}
\arguments{
\item{seu_path}{File path}

\item{corresponding_states_dictionary}{Parameter for corresponding states dictionary}

\item{large_clone_comparisons}{Parameter for large clone comparisons}

\item{numbat_rds_files}{File path}

\item{location}{Character string (default: "cis")}
}
\value{
Differential expression results
}
\description{
Perform differential expression analysis
}
