% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_48.R
\name{find_diffex_from_clustree}
\alias{find_diffex_from_clustree}
\title{Perform differential expression analysis}
\usage{
find_diffex_from_clustree(
  to_SCT_snn_res. = 1,
  to_clust = "1_10",
  sample_id,
  tumor_id,
  seu,
  mynb,
  ...
)
}
\arguments{
\item{to_SCT_snn_res.}{Parameter for to SCT snn res.}

\item{to_clust}{Character string (default: "1_10")}

\item{sample_id}{Parameter for sample id}

\item{tumor_id}{Parameter for tumor id}

\item{seu}{Seurat object}

\item{mynb}{Numbat object}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Perform differential expression analysis
}
