% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_52.R
\name{find_genes_by_arm}
\alias{find_genes_by_arm}
\title{Perform find genes by arm operation}
\usage{
find_genes_by_arm(gene_list)
}
\arguments{
\item{gene_list}{Gene names or identifiers}
}
\value{
Function result
}
\description{
Perform find genes by arm operation
}
