% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_43.R
\name{generate_plae_ref}
\alias{generate_plae_ref}
\title{Perform generate plae ref operation}
\usage{
generate_plae_ref(plae_seu_path = "data/plae_human_fetal_seu.rds")
}
\arguments{
\item{plae_seu_path}{File path}
}
\value{
Function result
}
\description{
Perform generate plae ref operation
}
