% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_10.R
\name{get_chr_arms}
\alias{get_chr_arms}
\title{Perform get chr arms operation}
\usage{
get_chr_arms(build = "hg19")
}
\arguments{
\item{build}{Character string (default: "hg19")}
}
\value{
Function result
}
\description{
Perform get chr arms operation
}
