% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seu_metadata_db.R
\name{get_cluster_markers}
\alias{get_cluster_markers}
\title{Retrieve top marker genes for one Seurat object}
\usage{
get_cluster_markers(filepath, top_n = 5, sqlite_path = DEFAULT_DB)
}
\arguments{
\item{filepath}{Path used as primary key in \code{seurat_objects}.}

\item{top_n}{Number of top markers to return per cluster.}

\item{sqlite_path}{Path to the SQLite database.}
}
\value{
A data frame with cluster, marker_rank, and gene_name columns.
}
\description{
Retrieve top marker genes for one Seurat object
}
