% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_23.R
\name{get_merged_metadata}
\alias{get_merged_metadata}
\title{Merge or join datasets}
\usage{
get_merged_metadata(meta_path)
}
\arguments{
\item{meta_path}{File path}
}
\value{
Data frame
}
\description{
Merge or join datasets
}
