% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_27.R
\name{gse_plot_from_clone_diffex}
\alias{gse_plot_from_clone_diffex}
\title{Create a plot visualization}
\usage{
gse_plot_from_clone_diffex(diffex_path)
}
\arguments{
\item{diffex_path}{File path}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
