% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_26.R
\name{gse_plot_from_cluster_diffex}
\alias{gse_plot_from_cluster_diffex}
\title{Create a plot visualization}
\usage{
gse_plot_from_cluster_diffex(diffex_path)
}
\arguments{
\item{diffex_path}{File path}

\item{mylabel}{Parameter for mylabel}

\item{plot_list}{Parameter for plot list}
}
\value{
ggplot2 plot object

Function result

ggplot2 plot object
}
\description{
Create a plot visualization
}
