% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_17.R
\name{heatmap_from_cluster_diffex}
\alias{heatmap_from_cluster_diffex}
\title{Perform differential expression analysis}
\usage{
heatmap_from_cluster_diffex(
  seu,
  enrichments,
  diffexes,
  myset = "HALLMARK_E2F_TARGETS",
  mycomparison = "g1_6 v. g1_0"
)
}
\arguments{
\item{seu}{Seurat object}

\item{enrichments}{Parameter for enrichments}

\item{diffexes}{Parameter for diffexes}

\item{myset}{Character string (default: "HALLMARK_E2F_TARGETS")}

\item{mycomparison}{Character string (default: "g1_6 v. g1_0")}
}
\value{
Differential expression results
}
\description{
Perform differential expression analysis
}
