% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_34.R
\name{inspect_oncoprints}
\alias{inspect_oncoprints}
\title{Perform inspect oncoprints operation}
\usage{
inspect_oncoprints(cis_comps, trans_comps, all_comps)
}
\arguments{
\item{cis_comps}{Parameter for cis comps}

\item{trans_comps}{Parameter for trans comps}

\item{all_comps}{Parameter for all comps}
}
\value{
List object
}
\description{
Perform inspect oncoprints operation
}
