% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_19.R
\name{label_enrichment_by_cis}
\alias{label_enrichment_by_cis}
\title{Perform enrichment analysis}
\usage{
label_enrichment_by_cis(mytable, mychrom = "01", myarm = "q")
}
\arguments{
\item{mytable}{Parameter for mytable}

\item{mychrom}{Character string (default: "01")}

\item{myarm}{Character string (default: "q")}
}
\value{
Enrichment analysis results
}
\description{
Perform enrichment analysis
}
