% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_16.R
\name{make_annoHighlight_from_consensus}
\alias{make_annoHighlight_from_consensus}
\title{Perform make annohighlight from consensus operation}
\usage{
make_annoHighlight_from_consensus(
  ideogramPlot,
  ploty,
  chrom,
  chromstart,
  chromend,
  fill,
  width,
  pg_width
)
}
\arguments{
\item{ideogramPlot}{Parameter for ideogramPlot}

\item{ploty}{Parameter for ploty}

\item{chrom}{Parameter for chrom}

\item{chromstart}{Parameter for chromstart}

\item{chromend}{Parameter for chromend}

\item{fill}{Parameter for fill}

\item{width}{Parameter for width}

\item{pg_width}{Parameter for pg width}
}
\value{
Function result
}
\description{
Perform make annohighlight from consensus operation
}
