% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_9.R
\name{make_cc_plot}
\alias{make_cc_plot}
\title{Create a plot visualization}
\usage{
make_cc_plot(seu, var_y = "clusters")
}
\arguments{
\item{seu}{Seurat object}

\item{var_y}{Character string (default: "clusters")}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
