% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions_1.R
\name{make_cell_cycle_plot}
\alias{make_cell_cycle_plot}
\title{Create a plot visualization}
\usage{
make_cell_cycle_plot(sample_id, myseus)
}
\arguments{
\item{sample_id}{Parameter for sample id}

\item{myseus}{Parameter for myseus}
}
\value{
Data frame
}
\description{
Create a plot visualization
}
