% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_2.R
\name{make_clone_comparison}
\alias{make_clone_comparison}
\title{Perform make clone comparison operation}
\usage{
make_clone_comparison(mysegs, comparison, seu, mynb, location = "cis")
}
\arguments{
\item{mysegs}{Parameter for mysegs}

\item{comparison}{Parameter for comparison}

\item{seu}{Seurat object}

\item{mynb}{Numbat object}

\item{location}{Character string (default: "cis")}
}
\value{
Function result
}
\description{
Perform make clone comparison operation
}
