% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_2.R
\name{make_clone_comparison_integrated}
\alias{make_clone_comparison_integrated}
\title{Perform make clone comparison integrated operation}
\usage{
make_clone_comparison_integrated(
  mysegs,
  comparison,
  seu,
  mynbs,
  location = "cis"
)
}
\arguments{
\item{mysegs}{Parameter for mysegs}

\item{comparison}{Parameter for comparison}

\item{seu}{Seurat object}

\item{mynbs}{Parameter for mynbs}

\item{location}{Character string (default: "cis")}

\item{mynb}{Numbat object}

\item{idents}{Cell identities or groups}
}
\value{
Function result

Extracted data elements
}
\description{
Perform make clone comparison integrated operation
}
