% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_2.R
\name{make_cluster_comparison}
\alias{make_cluster_comparison}
\title{Perform clustering analysis}
\usage{
make_cluster_comparison(
  mysegs,
  clone_comparison,
  seu,
  mynb,
  w_scna = NULL,
  wo_scna = NULL
)
}
\arguments{
\item{mysegs}{Parameter for mysegs}

\item{clone_comparison}{Parameter for clone comparison}

\item{seu}{Seurat object}

\item{mynb}{Numbat object}

\item{w_scna}{Parameter for w scna}

\item{wo_scna}{Parameter for wo scna}

\item{idents}{Cell identities or groups}
}
\value{
List object

Extracted data elements
}
\description{
Perform clustering analysis
}
