% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_16.R
\name{make_corresponding_states_dictionary}
\alias{make_corresponding_states_dictionary}
\title{Perform make corresponding states dictionary operation}
\usage{
make_corresponding_states_dictionary(my_tbl)
}
\arguments{
\item{my_tbl}{Parameter for my tbl}
}
\value{
Function result
}
\description{
Perform make corresponding states dictionary operation
}
