% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_28.R
\name{make_expression_heatmap_comparison}
\alias{make_expression_heatmap_comparison}
\title{Perform make expression heatmap comparison operation}
\usage{
make_expression_heatmap_comparison(
  large_numbat_pdfs,
  heatmaps,
  montage_path = "results/expr_heatmap_comparison.pdf"
)
}
\arguments{
\item{large_numbat_pdfs}{Parameter for large numbat pdfs}

\item{heatmaps}{Parameter for heatmaps}

\item{montage_path}{File path}
}
\value{
Function result
}
\description{
Perform make expression heatmap comparison operation
}
