% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_46.R
\name{make_faded_umap_plots}
\alias{make_faded_umap_plots}
\title{Create a plot visualization}
\usage{
make_faded_umap_plots(full_seu, retained_clones, group_by = "clusters")
}
\arguments{
\item{full_seu}{Parameter for full seu}

\item{retained_clones}{Parameter for retained clones}

\item{group_by}{Character string (default: "clusters")}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
