% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_6.R
\name{make_filtered_numbat_plots}
\alias{make_filtered_numbat_plots}
\title{Create a numbat-related plot visualization}
\usage{
make_filtered_numbat_plots(
  output_file,
  sample_id,
  myseus,
  mynbs,
  merged_metadata,
  myexpressions
)
}
\arguments{
\item{output_file}{File path}

\item{sample_id}{Parameter for sample id}

\item{myseus}{Parameter for myseus}

\item{mynbs}{Parameter for mynbs}

\item{merged_metadata}{Parameter for merged metadata}

\item{myexpressions}{Parameter for myexpressions}
}
\value{
ggplot2 plot object
}
\description{
Create a numbat-related plot visualization
}
