% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_19.R
\name{make_gistic_plot}
\alias{make_gistic_plot}
\title{Create a plot visualization}
\usage{
make_gistic_plot(cohort = "READ", fullname = "READ", sig_peaks)
}
\arguments{
\item{cohort}{Character string (default: "READ")}

\item{fullname}{Character string (default: "READ")}

\item{sig_peaks}{Parameter for sig peaks}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
