% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_15.R
\name{make_integrated_numbat_plots}
\alias{make_integrated_numbat_plots}
\title{Make integrated numbat plots}
\usage{
make_integrated_numbat_plots(seu_path, extension = "")
}
\arguments{
\item{extension}{}
}
\description{
Make integrated numbat plots
}
