% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_20.R
\name{make_leng_sce}
\alias{make_leng_sce}
\title{Perform make leng sce operation}
\usage{
make_leng_sce(
  tibble,
  column_to_rownames,
  dplyr,
  mutate,
  enframe,
  sample_id,
  SingleCellExperiment
)
}
\arguments{
\item{tibble}{Parameter for tibble}

\item{column_to_rownames}{Color specification}

\item{dplyr}{Parameter for dplyr}

\item{mutate}{Parameter for mutate}

\item{enframe}{Parameter for enframe}

\item{sample_id}{Parameter for sample id}

\item{SingleCellExperiment}{Cell identifiers or information}
}
\value{
Function result
}
\description{
Perform make leng sce operation
}
