% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_16.R
\name{make_numbat_heatmaps_old}
\alias{make_numbat_heatmaps_old}
\title{Title}
\usage{
make_numbat_heatmaps_old(
  numbat_rds_file,
  filter_expressions = NULL,
  cluster_dictionary,
  p_min = 0.9,
  line_width = 0.1,
  extension = ""
)
}
\description{
Title
}
