% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_22.R
\name{make_numbat_plot_files}
\alias{make_numbat_plot_files}
\title{Create a numbat-related plot visualization}
\usage{
make_numbat_plot_files(
  seu_path,
  numbat_rds_files,
  cluster_dictionary,
  filter_expressions = NULL,
  clone_simplifications = NULL,
  extension = ""
)
}
\arguments{
\item{seu_path}{File path}

\item{numbat_rds_files}{File path}

\item{cluster_dictionary}{Cluster information}

\item{filter_expressions}{Parameter for filter expressions}

\item{clone_simplifications}{Parameter for clone simplifications}

\item{extension}{Character string (default: "")}
}
\value{
ggplot2 plot object
}
\description{
Create a numbat-related plot visualization
}
