% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_3.R
\name{make_oncoprint_diffex}
\alias{make_oncoprint_diffex}
\title{Perform differential expression analysis}
\usage{
make_oncoprint_diffex(
  large_filter_expressions,
  cluster_dictionary,
  interesting_samples,
  cis_diffex_clones,
  trans_diffex_clones,
  all_diffex_clones,
  large_clone_comparisons,
  rb_scna_samples,
  by_cluster = FALSE,
  ...
)
}
\arguments{
\item{large_filter_expressions}{Parameter for large filter expressions}

\item{cluster_dictionary}{Cluster information}

\item{interesting_samples}{Parameter for interesting samples}

\item{cis_diffex_clones}{Parameter for cis diffex clones}

\item{trans_diffex_clones}{Parameter for trans diffex clones}

\item{all_diffex_clones}{Parameter for all diffex clones}

\item{large_clone_comparisons}{Parameter for large clone comparisons}

\item{rb_scna_samples}{Parameter for rb scna samples}

\item{by_cluster}{Cluster information}

\item{...}{Additional arguments passed to other functions}

\item{df}{Input data frame or dataset}

\item{n_slice}{Parameter for n slice}

\item{diffex_clones}{Parameter for diffex clones}

\item{selected_scna}{Character string (default: "1q+")}
}
\value{
Differential expression results

Filtered data

Differential expression results
}
\description{
Perform differential expression analysis
}
