% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_34.R
\name{make_oncoprint_diffex_unfiltered}
\alias{make_oncoprint_diffex_unfiltered}
\title{Filter data based on specified criteria}
\usage{
make_oncoprint_diffex_unfiltered(
  large_filter_expressions,
  cluster_dictionary,
  interesting_samples,
  cis_diffex_clones,
  trans_diffex_clones,
  large_clone_comparisons,
  ...
)
}
\arguments{
\item{large_filter_expressions}{Parameter for large filter expressions}

\item{cluster_dictionary}{Cluster information}

\item{interesting_samples}{Parameter for interesting samples}

\item{cis_diffex_clones}{Parameter for cis diffex clones}

\item{trans_diffex_clones}{Parameter for trans diffex clones}

\item{large_clone_comparisons}{Parameter for large clone comparisons}

\item{...}{Additional arguments passed to other functions}
}
\value{
List object
}
\description{
Filter data based on specified criteria
}
