% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_34.R
\name{make_oncoprint_plots}
\alias{make_oncoprint_plots}
\title{Create a plot visualization}
\usage{
make_oncoprint_plots(
  comps,
  clone_trees,
  oncoprint_settings,
  label = "_by_clone",
  ...
)
}
\arguments{
\item{comps}{Parameter for comps}

\item{clone_trees}{Parameter for clone trees}

\item{oncoprint_settings}{Parameter for oncoprint settings}

\item{label}{Character string (default: "_by_clone")}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
