% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_44.R
\name{make_pairwise_plots}
\alias{make_pairwise_plots}
\title{Create a plot visualization}
\usage{
make_pairwise_plots(
  clone_set,
  clone_comparison,
  seu,
  sample_id,
  var_y = "clusters"
)
}
\arguments{
\item{clone_set}{Parameter for clone set}

\item{clone_comparison}{Parameter for clone comparison}

\item{seu}{Seurat object}

\item{sample_id}{Parameter for sample id}

\item{var_y}{Character string (default: "clusters")}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
