% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_28.R
\name{make_pdf_montages}
\alias{make_pdf_montages}
\title{Perform make pdf montages operation}
\usage{
make_pdf_montages(plot_files, heatmaps, tile = "6")
}
\arguments{
\item{plot_files}{File path}

\item{heatmaps}{Parameter for heatmaps}

\item{tile}{Character string (default: "6")}
}
\value{
Function result
}
\description{
Perform make pdf montages operation
}
