% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_21.R
\name{make_pearls_plot}
\alias{make_pearls_plot}
\title{Create a plot visualization}
\usage{
make_pearls_plot(df, y_setting = 4, var_y = "clusters")
}
\arguments{
\item{df}{Input data frame or dataset}

\item{y_setting}{Parameter for y setting}

\item{var_y}{Character string (default: "clusters")}

\item{x}{Parameter for x}

\item{n}{Parameter for n}
}
\value{
ggplot2 plot object

Function result
}
\description{
Create a plot visualization
}
