% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_16.R
\name{make_rb_scna_ideograms}
\alias{make_rb_scna_ideograms}
\title{Perform make rb scna ideograms operation}
\usage{
make_rb_scna_ideograms(nb_path, midline_threshold = 0.4, suffix = "")
}
\arguments{
\item{nb_path}{File path}

\item{midline_threshold}{Threshold value for filtering}

\item{suffix}{Optional suffix inserted before the karyogram filename stem}
}
\value{
ggplot2 plot object
}
\description{
Perform make rb scna ideograms operation
}
