% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_16.R
\name{make_rb_scna_ideograms_old}
\alias{make_rb_scna_ideograms_old}
\title{Perform make rb scna ideograms old operation}
\usage{
make_rb_scna_ideograms_old(consensus_file, suffix = "")
}
\arguments{
\item{consensus_file}{File path}

\item{suffix}{Optional suffix inserted before the karyogram filename stem}
}
\value{
ggplot2 plot object
}
\description{
Perform make rb scna ideograms old operation
}
