% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_12.R
\name{make_table_s01}
\alias{make_table_s01}
\title{Perform make table s01 operation}
\usage{
make_table_s01(study_cell_stats, path = "doc/table_s01.csv")
}
\arguments{
\item{study_cell_stats}{Cell identifiers or information}

\item{path}{File path}
}
\value{
Function result
}
\description{
Perform make table s01 operation
}
