% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_13.R
\name{make_table_s02}
\alias{make_table_s02}
\title{Perform make table s02 operation}
\usage{
make_table_s02(meta_path = "data/metadata.tsv")
}
\arguments{
\item{meta_path}{File path}
}
\value{
Function result
}
\description{
Perform make table s02 operation
}
