% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_13.R
\name{make_table_s03}
\alias{make_table_s03}
\title{Perform make table s03 operation}
\usage{
make_table_s03(
  cluster_dictionary,
  table_path = "results/table_s03.csv",
  sqlite_path = "batch_hashes.sqlite"
)
}
\arguments{
\item{cluster_dictionary}{Cluster information}

\item{table_path}{File path}

\item{sqlite_path}{Path to Seurat metadata SQLite database}
}
\value{
Function result
}
\description{
Perform make table s03 operation
}
