% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_10.R
\name{make_table_s05}
\alias{make_table_s05}
\title{Perform make table s05 operation}
\usage{
make_table_s05(table_path = "results/table_s05.xlsx")
}
\arguments{
\item{table_path}{File path}
}
\value{
Function result
}
\description{
Perform make table s05 operation
}
