% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_19.R
\name{make_table_s07}
\alias{make_table_s07}
\title{Perform make table s07 operation}
\usage{
make_table_s07(
  seu_path = "output/seurat/integrated_1q/integrated_seu_1q_complete.rds",
  table_path = "results/table_s07.csv"
)
}
\arguments{
\item{seu_path}{File path}

\item{table_path}{File path}
}
\value{
Function result
}
\description{
Perform make table s07 operation
}
