% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_13.R
\name{make_table_s08}
\alias{make_table_s08}
\title{Perform make table s08 operation}
\usage{
make_table_s08(table_path = "results/table_s04.csv")
}
\arguments{
\item{table_path}{File path}
}
\value{
Function result
}
\description{
Perform make table s08 operation
}
