% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_32.R
\name{make_volcano_diffex_clones}
\alias{make_volcano_diffex_clones}
\title{Perform differential expression analysis}
\usage{
make_volcano_diffex_clones(
  cluster_diffex_clones,
  cluster_pdf = "results/diffex_bw_clones_per_cluster_large.pdf",
  total_diffex_clones,
  total_pdf = "results/straight_diffex_bw_clones_large.pdf"
)
}
\arguments{
\item{cluster_diffex_clones}{Cluster information}

\item{cluster_pdf}{Cluster information}

\item{total_diffex_clones}{Parameter for total diffex clones}

\item{total_pdf}{Character string (default: "results/straight_diffex_bw_clones_large.pdf")}
}
\value{
Data frame
}
\description{
Perform differential expression analysis
}
