% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_51.R
\name{merge_orders}
\alias{merge_orders}
\title{Merge or join datasets}
\usage{
merge_orders(...)
}
\arguments{
\item{...}{Additional arguments passed to other functions}
}
\value{
Function result
}
\description{
Merge or join datasets
}
