% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{not_sure_what_this_does}
\alias{not_sure_what_this_does}
\title{Assemble supplementary figure s07 (clone distribution + cc plots)}
\usage{
not_sure_what_this_does(
  seu_list,
  cluster_orders,
  plot_path = "results/fig_s07.pdf"
)
}
\arguments{
\item{seu_list}{Parameter for seu list}

\item{cluster_orders}{Cluster information}

\item{plot_path}{File path}
}
\value{
File path of assembled PDF
}
\description{
Assemble supplementary figure s07 (clone distribution + cc plots)
}
