% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions_1.R
\name{ora_analysis}
\alias{ora_analysis}
\title{Perform ora analysis operation}
\usage{
ora_analysis(seu, clusters = "gene_snn_res.0.2", only_unique_terms = FALSE)
}
\arguments{
\item{seu}{Seurat object}

\item{clusters}{Cluster information}

\item{only_unique_terms}{Logical flag (default: FALSE)}

\item{df}{Input data frame or dataset}

\item{fold_change_col}{Color specification}

\item{cluster_name}{Cluster information}

\item{ora_results}{Parameter for ora results}

\item{terms_list}{Parameter for terms list}
}
\value{
Function result

Enrichment analysis results

Filtered data
}
\description{
Perform ora analysis operation
}
