% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_41.R
\name{ora_effect_of_regression}
\alias{ora_effect_of_regression}
\title{Perform ora effect of regression operation}
\usage{
ora_effect_of_regression(
  filtered_seu_path,
  regressed_seu_path,
  resolution = "0.4"
)
}
\arguments{
\item{filtered_seu_path}{File path}

\item{regressed_seu_path}{File path}

\item{resolution}{Character string (default: "0.4")}
}
\value{
ggplot2 plot object
}
\description{
Perform ora effect of regression operation
}
