% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_10.R
\name{paired_plots_to_pptx}
\alias{paired_plots_to_pptx}
\title{Create a plot visualization}
\usage{
paired_plots_to_pptx(plots1, plots2, file, ...)
}
\arguments{
\item{plots1}{Parameter for plots1}

\item{plots2}{Parameter for plots2}

\item{file}{File path}

\item{...}{Additional arguments passed to other functions}

\item{ppt}{Parameter for ppt}

\item{plot1}{Parameter for plot1}

\item{plot2}{Parameter for plot2}

\item{h1}{Character string (default: "heading 1")}

\item{h2}{Character string (default: "heading 2")}
}
\value{
ggplot2 plot object

ggplot2 plot object
}
\description{
Create a plot visualization
}
