% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_53.R
\name{plot_cc_space_plot}
\alias{plot_cc_space_plot}
\title{Create a plot visualization}
\usage{
plot_cc_space_plot(
  seu_path = "output/seurat/SRX11133594_filtered_seu.rds",
  facet = FALSE,
  group_by = "clusters",
  color_by = "clusters"
)
}
\arguments{
\item{seu_path}{File path}

\item{facet}{Logical flag (default: FALSE)}

\item{group_by}{Character string (default: "clusters")}

\item{color_by}{Color specification}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
