% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_43.R
\name{plot_celltype_predictions}
\alias{plot_celltype_predictions}
\title{Create a plot visualization}
\usage{
plot_celltype_predictions(
  seu_path,
  sample_id,
  plae_ref = NULL,
  group.by = "SCT_snn_res.0.4",
  query_genes = NULL
)
}
\arguments{
\item{seu_path}{File path}

\item{sample_id}{Parameter for sample id}

\item{plae_ref}{Parameter for plae ref}

\item{group.by}{Character string (default: "SCT_snn_res.0.4")}

\item{query_genes}{Gene names or identifiers}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
