% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_7.R
\name{plot_cluster_markers_by_cell_type}
\alias{plot_cluster_markers_by_cell_type}
\title{Create a plot visualization}
\usage{
plot_cluster_markers_by_cell_type(seu, checked_cluster_markers)
}
\arguments{
\item{seu}{Seurat object}

\item{checked_cluster_markers}{Cluster information}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
