% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_45.R
\name{plot_clustree_per_comparison}
\alias{plot_clustree_per_comparison}
\title{Create a plot visualization}
\usage{
plot_clustree_per_comparison(
  mylabel,
  clustree_plot,
  speckle_proportions,
  sample_id
)
}
\arguments{
\item{mylabel}{Parameter for mylabel}

\item{clustree_plot}{Parameter for clustree plot}

\item{speckle_proportions}{Parameter for speckle proportions}

\item{sample_id}{Parameter for sample id}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
