% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_9.R
\name{plot_corresponding_clusters_diffex_heatmaps}
\alias{plot_corresponding_clusters_diffex_heatmaps}
\title{Create a heatmap visualization}
\usage{
plot_corresponding_clusters_diffex_heatmaps(
  diffex_list,
  seu_path,
  corresponding_states_dictionary,
  large_clone_comparisons,
  numbat_rds_files = numbat_rds_files,
  location = "all"
)
}
\arguments{
\item{diffex_list}{Parameter for diffex list}

\item{seu_path}{File path}

\item{corresponding_states_dictionary}{Parameter for corresponding states dictionary}

\item{large_clone_comparisons}{Parameter for large clone comparisons}

\item{numbat_rds_files}{File path}

\item{location}{Character string (default: "all")}
}
\value{
ggplot2 plot object
}
\description{
Create a heatmap visualization
}
