% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_9.R
\name{plot_corresponding_clusters_diffex_volcanos}
\alias{plot_corresponding_clusters_diffex_volcanos}
\title{Generate a volcano plot for differential expression data}
\usage{
plot_corresponding_clusters_diffex_volcanos(diffex_list, seu_path, ...)
}
\arguments{
\item{diffex_list}{Parameter for diffex list}

\item{seu_path}{File path}

\item{...}{Additional arguments passed to other functions}
}
\value{
Data frame
}
\description{
Generate a volcano plot for differential expression data
}
