% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_11.R
\name{plot_corresponding_enrichment}
\alias{plot_corresponding_enrichment}
\title{Create a plot visualization}
\usage{
plot_corresponding_enrichment(sample_diffex_list, filename, ...)
}
\arguments{
\item{sample_diffex_list}{Parameter for sample diffex list}

\item{filename}{File path}

\item{...}{Additional arguments passed to other functions}

\item{enrichment_result}{Parameter for enrichment result}
}
\value{
ggplot2 plot object

Filtered data
}
\description{
Create a plot visualization
}
