% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_15.R
\name{plot_diffex_genes_on_split_integrated_seu}
\alias{plot_diffex_genes_on_split_integrated_seu}
\title{Create a plot visualization}
\usage{
plot_diffex_genes_on_split_integrated_seu(
  plot_path = "results/fig_07d.pdf",
  integrated_seu_path = "output/seurat/integrated_1q/integrated_seu_1q_complete.rds",
  direction = "up",
  p_val_adj_threshold = 0.05,
  recurrence_threshold = 3,
  n_genes = 10,
  ident.1 = "w_scna",
  ident.2 = "wo_scna",
  ...
)
}
\arguments{
\item{plot_path}{File path}

\item{integrated_seu_path}{File path}

\item{direction}{Character string (default: "up")}

\item{p_val_adj_threshold}{Threshold value for filtering}

\item{recurrence_threshold}{Threshold value for filtering}

\item{n_genes}{Gene names or identifiers}

\item{ident.1}{Cell identities or groups}

\item{ident.2}{Cell identities or groups}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
