% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_2.R
\name{plot_distribution_of_clones_across_clusters}
\alias{plot_distribution_of_clones_across_clusters}
\title{plot distribution of clones across clusters}
\usage{
plot_distribution_of_clones_across_clusters(
  seu,
  seu_name,
  var_x = "scna",
  var_y = "SCT_snn_res.0.6",
  plot_type = c("both", "clone", "cluster"),
  avg_line = NULL,
  signif = FALSE,
  integrated = FALSE
)
}
\arguments{
\item{clusters}{}
}
\description{
plot distribution of clones across clusters
}
