% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_21.R
\name{plot_distribution_of_clones_pearls}
\alias{plot_distribution_of_clones_pearls}
\title{Create a plot visualization}
\usage{
plot_distribution_of_clones_pearls(
  seu,
  seu_name,
  var_x = "scna",
  var_y = "SCT_snn_res.0.6",
  avg_line = NULL
)
}
\arguments{
\item{seu}{Seurat object}

\item{seu_name}{Parameter for seu name}

\item{var_x}{Character string (default: "scna")}

\item{var_y}{Character string (default: "SCT_snn_res.0.6")}

\item{avg_line}{Parameter for avg line}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
