% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_40.R
\name{plot_effect_of_filtering}
\alias{plot_effect_of_filtering}
\title{Create a plot visualization}
\usage{
plot_effect_of_filtering(
  unfiltered_seu_path,
  filtered_seu_path = NULL,
  group.by = "gene_snn_res.0.2",
  cluster_dictionary,
  mito_threshold = c(5, 10),
  nCount_threshold = c(500, 1000),
  nFeature_threshold = c(500, 1000),
  low_hypoxia_seu_path = NULL,
  high_hypoxia_seu_path = NULL,
  plot_path = NULL
)
}
\arguments{
\item{unfiltered_seu_path}{File path}

\item{filtered_seu_path}{File path}

\item{group.by}{Character string (default: "gene_snn_res.0.2")}

\item{cluster_dictionary}{Cluster information}

\item{plot_path}{File path}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
