% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_40.R
\name{plot_effect_of_filtering_old}
\alias{plot_effect_of_filtering_old}
\title{Create a plot visualization}
\usage{
plot_effect_of_filtering_old(
  unfiltered_seu_path,
  filtered_seu_path,
  group.by = "SCT_snn_res.0.6"
)
}
\arguments{
\item{unfiltered_seu_path}{File path}

\item{filtered_seu_path}{File path}

\item{group.by}{Character string (default: "SCT_snn_res.0.6")}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
