% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_40.R
\name{plot_effect_of_regression}
\alias{plot_effect_of_regression}
\title{Create a plot visualization}
\usage{
plot_effect_of_regression(
  filtered_seu_path,
  regressed_seu_path,
  resolution = 0.4,
  filter_dropped_cluster = NULL,
  regress_dropped_cluster = NULL,
  n_features = 2,
  ...
)
}
\arguments{
\item{filtered_seu_path}{File path}

\item{regressed_seu_path}{File path}

\item{resolution}{Parameter for resolution}

\item{filter_dropped_cluster}{Cluster information}

\item{regress_dropped_cluster}{Cluster information}

\item{n_features}{Parameter for n features}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
