% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_40.R
\name{plot_effect_of_regression_old}
\alias{plot_effect_of_regression_old}
\title{Create a plot visualization}
\usage{
plot_effect_of_regression_old(
  filtered_seu_path,
  regressed_seu_path,
  resolution = "0.4",
  group.by = "SCT_snn_res.0.6"
)
}
\arguments{
\item{filtered_seu_path}{File path}

\item{regressed_seu_path}{File path}

\item{resolution}{Character string (default: "0.4")}

\item{group.by}{Character string (default: "SCT_snn_res.0.6")}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
