% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions_1.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Create a plot visualization}
\usage{
plot_enrichment(
  gse,
  p_val_cutoff = NULL,
  signed = TRUE,
  analysis_method = c("gsea", "ora"),
  result_slot = "result",
  showCategory = 10
)
}
\arguments{
\item{gse}{Parameter for gse}

\item{showCategory}{Parameter for showCategory}
}
\value{
ggplot2 plot object

ggplot2 plot object
}
\description{
Create a plot visualization
}
