% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_38.R
\name{plot_enrichment_per_scna}
\alias{plot_enrichment_per_scna}
\title{Create a plot visualization}
\usage{
plot_enrichment_per_scna(enrichment_table_list, scna, input_plot_file)
}
\arguments{
\item{enrichment_table_list}{Parameter for enrichment table list}

\item{scna}{Parameter for scna}

\item{input_plot_file}{File path}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
