% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_38.R
\name{plot_enrichment_recurrence_by_cluster}
\alias{plot_enrichment_recurrence_by_cluster}
\title{Create a plot visualization}
\usage{
plot_enrichment_recurrence_by_cluster(
  enrichment_tables,
  num_recur = 2,
  mytitle = "",
  n_slice = 10,
  by_cluster = TRUE,
  pvalueCutoff = 0.3
)
}
\arguments{
\item{enrichment_tables}{Parameter for enrichment tables}

\item{num_recur}{Parameter for num recur}

\item{mytitle}{Plot title}

\item{n_slice}{Parameter for n slice}

\item{by_cluster}{Cluster information}

\item{pvalueCutoff}{Threshold value for filtering}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
