% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_5.R
\name{plot_feature_across_seus}
\alias{plot_feature_across_seus}
\title{Generate a feature plot}
\usage{
plot_feature_across_seus(seu, seu_title, myvar)
}
\arguments{
\item{seu}{Seurat object}

\item{seu_title}{Plot title}

\item{myvar}{Parameter for myvar}
}
\value{
ggplot2 plot object
}
\description{
Generate a feature plot
}
