% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_13.R
\name{plot_fig_03}
\alias{plot_fig_03}
\title{Create a plot visualization}
\usage{
plot_fig_03(cluster_orders, plot_path = "results/fig_03.pdf")
}
\arguments{
\item{cluster_orders}{Cluster information}

\item{plot_path}{File path}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
