% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_11.R
\name{plot_fig_03_09}
\alias{plot_fig_03_09}
\title{Create a plot visualization}
\usage{
plot_fig_03_09(sample_diffex_list, filename, ...)
}
\arguments{
\item{sample_diffex_list}{Parameter for sample diffex list}

\item{filename}{File path}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
