% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R, R/plot_functions_14.R
\name{plot_fig_04_05}
\alias{plot_fig_04_05}
\title{Create a plot visualization}
\usage{
plot_fig_04_05(
  seu_paths,
  integrated_enrichment,
  plot_path = "results/fig_04.pdf",
  integrated_seu_paths,
  ...
)

plot_fig_04_05(
  seu_paths,
  integrated_enrichment,
  plot_path = "results/fig_04.pdf",
  integrated_seu_paths,
  ...
)
}
\arguments{
\item{seu_paths}{File path}

\item{integrated_enrichment}{Parameter for integrated enrichment}

\item{plot_path}{File path}

\item{integrated_seu_paths}{File path}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object

ggplot2 plot object
}
\description{
Create a plot visualization

Create a plot visualization
}
