% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R, R/plot_functions_13.R
\name{plot_fig_04_07}
\alias{plot_fig_04_07}
\title{Create a plot visualization}
\usage{
plot_fig_04_07(cluster_orders, plot_path = "results/fig_04_07.pdf")

plot_fig_04_07(cluster_orders, plot_path = "results/fig_04_07.pdf")
}
\arguments{
\item{cluster_orders}{Cluster information}

\item{plot_path}{File path}
}
\value{
ggplot2 plot object

ggplot2 plot object
}
\description{
Create a plot visualization

Create a plot visualization
}
