% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{plot_fig_04_afterall}
\alias{plot_fig_04_afterall}
\title{Create a plot visualization}
\usage{
plot_fig_04_afterall(cluster_orders, plot_path = "results/fig_02afterall.pdf")
}
\arguments{
\item{cluster_orders}{Cluster information}

\item{plot_path}{File path}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
