% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_15.R
\name{plot_fig_08d}
\alias{plot_fig_08d}
\title{Create a plot visualization}
\usage{
plot_fig_08d(
  plot_path = "results/fig_08d.pdf",
  integrated_seu_path = "output/seurat/integrated_16q/integrated_seu_16q_complete.rds",
  direction = "down",
  p_val_adj_threshold = 0.1,
  recurrence_threshold = 2,
  n_genes = 10,
  ident.1 = "16q-",
  ident.2 = ".diploid",
  ...
)
}
\arguments{
\item{plot_path}{File path}

\item{integrated_seu_path}{File path}

\item{direction}{Character string (default: "down")}

\item{p_val_adj_threshold}{Threshold value for filtering}

\item{recurrence_threshold}{Threshold value for filtering}

\item{n_genes}{Gene names or identifiers}

\item{ident.1}{Cell identities or groups}

\item{ident.2}{Cell identities or groups}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
