% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{plot_fig_09_10}
\alias{plot_fig_09_10}
\title{Create a plot visualization}
\usage{
plot_fig_09_10(
  seu_subset,
  corresponding_seus,
  corresponding_clusters_diffex,
  corresponding_clusters_enrichments,
  plot_path = "result/dotplot_recurre.pdf",
  widths = rep(8, 3),
  heights = rep(12, 3),
  common_seus = NULL,
  ...
)
}
\arguments{
\item{seu_subset}{Parameter for seu subset}

\item{corresponding_seus}{Parameter for corresponding seus}

\item{corresponding_clusters_diffex}{Parameter for corresponding clusters diffex}

\item{corresponding_clusters_enrichments}{Parameter for corresponding clusters enrichments}

\item{plot_path}{File path}

\item{widths}{Parameter for widths}

\item{heights}{Parameter for heights}

\item{common_seus}{Parameter for common seus}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
