% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{plot_fig_s06a}
\alias{plot_fig_s06a}
\title{Create supplementary figure s06a (SCNA ideograms for all samples)}
\usage{
plot_fig_s06a(
  plot_path = "results/fig_s06a.pdf",
  table_path = "results/table_s12.xlsx"
)
}
\arguments{
\item{plot_path}{File path}

\item{table_path}{File path}
}
\value{
List of plot and table paths
}
\description{
Create supplementary figure s06a (SCNA ideograms for all samples)
}
