% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R, R/plot_functions_18.R
\name{plot_fig_s08}
\alias{plot_fig_s08}
\title{Create supplementary figure s08 (1q diffex per cluster)}
\usage{
plot_fig_s08(
  plot_path = "results/fig_s08.pdf",
  table_path = "results/table_s22.csv"
)

plot_fig_s08(
  plot_path = "results/fig_s08.pdf",
  table_path = "results/table_s22.csv"
)
}
\arguments{
\item{plot_path}{File path}

\item{table_path}{File path}
}
\value{
List of plot and table paths

ggplot2 plot object
}
\description{
Create supplementary figure s08 (1q diffex per cluster)

Create a plot visualization
}
