% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R, R/plot_functions_15.R
\name{plot_fig_s09}
\alias{plot_fig_s09}
\title{Create supplementary figure s09 (1q cluster comparisons: enrichments)}
\usage{
plot_fig_s09()

plot_fig_s09()
}
\value{
List of diffex table path and enrichment plot path

ggplot2 plot object
}
\description{
Create supplementary figure s09 (1q cluster comparisons: enrichments)

Create a plot visualization
}
