% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_18.R
\name{plot_fig_s10}
\alias{plot_fig_s10}
\title{Create a plot visualization}
\usage{
plot_fig_s10(
  plot_path = "results/fig_s10.pdf",
  table_path = "results/table_s23.csv"
)
}
\arguments{
\item{plot_path}{File path}

\item{table_path}{File path}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
