% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R, R/plot_functions_18.R
\name{plot_fig_s20}
\alias{plot_fig_s20}
\title{Create supplementary figure s20 (2p diffex per cluster)}
\usage{
plot_fig_s20()

plot_fig_s20()
}
\value{
List of plot and table paths

ggplot2 plot object
}
\description{
Create supplementary figure s20 (2p diffex per cluster)

Create a plot visualization
}
