% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_20.R
\name{plot_fig_s25}
\alias{plot_fig_s25}
\title{Create a plot visualization}
\usage{
plot_fig_s25(
  seu_path = "output/seurat/integrated_1q/integrated_seu_1q_complete.rds",
  subtype_markers
)
}
\arguments{
\item{seu_path}{File path}

\item{subtype_markers}{Parameter for subtype markers}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
