% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_44.R
\name{plot_filtering_timeline}
\alias{plot_filtering_timeline}
\title{Create a plot visualization}
\usage{
plot_filtering_timeline(
  all_cells_meta,
  scna_meta,
  qc_meta,
  cell_type_meta,
  sample_id
)
}
\arguments{
\item{all_cells_meta}{Cell identifiers or information}

\item{scna_meta}{Parameter for scna meta}

\item{qc_meta}{Parameter for qc meta}

\item{cell_type_meta}{Cell identifiers or information}

\item{sample_id}{Parameter for sample id}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
