% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_12.R
\name{plot_integrated_1q_fig}
\alias{plot_integrated_1q_fig}
\title{Create a plot visualization}
\usage{
plot_integrated_1q_fig(cluster_orders, plot_path = "results/fig_02.pdf")
}
\arguments{
\item{cluster_orders}{Cluster information}

\item{plot_path}{File path}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
