% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_12.R
\name{plot_integrated_1q_fig_low_hypoxia}
\alias{plot_integrated_1q_fig_low_hypoxia}
\title{Create a plot visualization}
\usage{
plot_integrated_1q_fig_low_hypoxia(
  seu_path = "output/seurat/integrated_1q/integrated_seu_1q_complete.rds",
  cluster_orders,
  plot_path = tempfile(tmpdir = "results", fileext = ".pdf")
)
}
\arguments{
\item{cluster_orders}{Cluster information}

\item{plot_path}{File path}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
