% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_24.R
\name{plot_markers_in_sample}
\alias{plot_markers_in_sample}
\title{Create a plot visualization}
\usage{
plot_markers_in_sample(
  seu_path,
  mymarkers,
  plot_type = FeaturePlot,
  group_by = group_by
)
}
\arguments{
\item{seu_path}{File path}

\item{mymarkers}{Parameter for mymarkers}

\item{plot_type}{Parameter for plot type}

\item{group_by}{Parameter for group by}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
