% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_9.R
\name{plot_merged_heatmap}
\alias{plot_merged_heatmap}
\title{Create a heatmap visualization}
\usage{
plot_merged_heatmap(
  seu_path = "output/seurat/merged_1q_filtered_seu.rds",
  child_seu_paths = debranched_seus_1q
)
}
\arguments{
\item{seu_path}{File path}

\item{child_seu_paths}{File path}
}
\value{
ggplot2 plot object
}
\description{
Create a heatmap visualization
}
