% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_numbat}
\alias{plot_numbat}
\title{Plot numbat heatmap}
\usage{
plot_numbat(nb, myseu, myannot, mytitle, sort_by = "scna", ...)
}
\arguments{
\item{nb}{Numbat object}

\item{myseu}{Seurat object}

\item{myannot}{Annotation data frame}

\item{mytitle}{Plot title}

\item{sort_by}{Sorting variable}

\item{...}{Additional arguments}
}
\value{
ggplot object
}
\description{
Plot numbat heatmap
}
