% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_numbat_w_phylo}
\alias{plot_numbat_w_phylo}
\title{Plot numbat heatmap with phylogeny}
\usage{
plot_numbat_w_phylo(nb, myseu, myannot, mytitle, ...)
}
\arguments{
\item{nb}{Numbat object}

\item{myseu}{Seurat object}

\item{myannot}{Annotation data frame}

\item{mytitle}{Plot title}

\item{...}{Additional arguments}
}
\value{
ggplot object
}
\description{
Plot numbat heatmap with phylogeny
}
