% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions_2.R
\name{plot_pcnv_by_reads}
\alias{plot_pcnv_by_reads}
\title{Create a plot visualization}
\usage{
plot_pcnv_by_reads(tbl, sample_id)
}
\arguments{
\item{tbl}{Parameter for tbl}

\item{sample_id}{Parameter for sample id}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
