% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_43.R
\name{plot_phase_distribution_by_scna}
\alias{plot_phase_distribution_by_scna}
\title{Create a plot visualization}
\usage{
plot_phase_distribution_by_scna(seu_path, seu_name)
}
\arguments{
\item{seu_path}{File path}

\item{seu_name}{Parameter for seu name}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
