% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_53.R
\name{plot_phase_wo_arm}
\alias{plot_phase_wo_arm}
\title{Create a plot visualization}
\usage{
plot_phase_wo_arm(seu_path, pdf_path = NULL)
}
\arguments{
\item{seu_path}{File path}

\item{pdf_path}{File path}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
