% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_25.R
\name{plot_putative_marker_across_samples}
\alias{plot_putative_marker_across_samples}
\title{Create a plot visualization}
\usage{
plot_putative_marker_across_samples(
  mymarkers,
  seu_paths,
  plot_type = FeaturePlot,
  group_by = "gene_snn_res.0.2",
  extension = "_filtered"
)
}
\arguments{
\item{mymarkers}{Parameter for mymarkers}

\item{seu_paths}{File path}

\item{plot_type}{Parameter for plot type}

\item{group_by}{Character string (default: "gene_snn_res.0.2")}

\item{extension}{Character string (default: "_filtered")}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
