% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_7.R
\name{plot_rb_subtype_expression}
\alias{plot_rb_subtype_expression}
\title{Create a plot visualization}
\usage{
plot_rb_subtype_expression(seu, seu_name, subtype_hallmarks)
}
\arguments{
\item{seu}{Seurat object}

\item{seu_name}{Parameter for seu name}

\item{subtype_hallmarks}{Parameter for subtype hallmarks}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
