% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_34.R
\name{plot_recurrence}
\alias{plot_recurrence}
\title{Create a plot visualization}
\usage{
plot_recurrence(
  diffex_input,
  scna_of_interest,
  segment_region = "cis",
  oncoprint_settings,
  clone_trees,
  n_genes = 20
)
}
\arguments{
\item{diffex_input}{Parameter for diffex input}

\item{scna_of_interest}{Parameter for scna of interest}

\item{segment_region}{Character string (default: "cis")}

\item{oncoprint_settings}{Parameter for oncoprint settings}

\item{clone_trees}{Parameter for clone trees}

\item{n_genes}{Gene names or identifiers}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
