% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_36.R
\name{plot_scna_violins}
\alias{plot_scna_violins}
\title{Create a plot visualization}
\usage{
plot_scna_violins(., readRDS, FeaturePlot, .x, .y, VlnPlot)
}
\arguments{
\item{.}{Parameter for .}

\item{readRDS}{Parameter for readRDS}

\item{FeaturePlot}{Parameter for FeaturePlot}

\item{.x}{Parameter for .x}

\item{.y}{Parameter for .y}

\item{VlnPlot}{Parameter for VlnPlot}

\item{sample_ids}{Parameter for sample ids}

\item{features}{Parameter for features}
}
\value{
ggplot2 plot object

ggplot2 plot object
}
\description{
Create a plot visualization
}
