% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R
\name{plot_seu_clusters_and_markers}
\alias{plot_seu_clusters_and_markers}
\title{Plot DimPlot + marker dotplot for a sample}
\usage{
plot_seu_clusters_and_markers(
  seu_path,
  cluster_order,
  phase_levels = c("pm", "g1", "g1_s", "s", "s_g2", "g2", "g2_m", "hsp", "hypoxia",
    "other", "s_star")
)
}
\arguments{
\item{seu_path}{File path to filtered Seurat object RDS}

\item{cluster_order}{Named list of cluster order data frames}

\item{phase_levels}{Ordered vector of phase level labels}
}
\value{
File path of output PDF
}
\description{
Plot DimPlot + marker dotplot for a sample
}
