% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_8.R
\name{plot_seu_gene_heatmap}
\alias{plot_seu_gene_heatmap}
\title{Create a heatmap visualization}
\usage{
plot_seu_gene_heatmap(seu_path, large_clone_comparisons, scna_of_interest, ...)
}
\arguments{
\item{seu_path}{File path}

\item{large_clone_comparisons}{Parameter for large clone comparisons}

\item{scna_of_interest}{Parameter for scna of interest}

\item{...}{Additional arguments passed to other functions}
}
\value{
Data frame
}
\description{
Create a heatmap visualization
}
