% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_46.R
\name{plot_seu_marker_heatmap_all_resolutions}
\alias{plot_seu_marker_heatmap_all_resolutions}
\title{Create a heatmap visualization}
\usage{
plot_seu_marker_heatmap_all_resolutions(
  seu_path = NULL,
  nb_paths = NULL,
  clone_simplifications = NULL,
  cluster_orders = NULL
)
}
\arguments{
\item{seu_path}{File path}

\item{nb_paths}{File path}

\item{clone_simplifications}{Parameter for clone simplifications}

\item{cluster_orders}{Cluster information}
}
\value{
ggplot2 plot object
}
\description{
Create a heatmap visualization
}
