% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_6.R
\name{plot_several_diffex_clones_in_phase}
\alias{plot_several_diffex_clones_in_phase}
\title{Title}
\usage{
plot_several_diffex_clones_in_phase(
  enrichments,
  scna_of_interest = "2p",
  phase = "g1"
)
}
\description{
Title
}
