% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_19.R
\name{plot_tcga_gistic}
\alias{plot_tcga_gistic}
\title{Create a plot visualization}
\usage{
plot_tcga_gistic(plot_path = "results/plot_gistic.pdf")
}
\arguments{
\item{plot_path}{File path}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
