% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R, R/plot_functions_1.R
\name{plot_variability_at_SCNA}
\alias{plot_variability_at_SCNA}
\title{Plot variability at SCNA}
\usage{
plot_variability_at_SCNA(phylo_plot_output, chrom = "1", p_min = 0.9)

plot_variability_at_SCNA(phylo_plot_output, chrom = "1", p_min = 0.9)
}
\arguments{
\item{phylo_plot_output}{Parameter for phylo plot output}

\item{chrom}{Character string (default: "1")}

\item{p_min}{Parameter for p min}
}
\value{
ggplot object

ggplot2 plot object
}
\description{
Plot variability at SCNA

Create a plot visualization
}
