% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_13.R
\name{plot_venn_w_genes}
\alias{plot_venn_w_genes}
\title{Create a plot visualization}
\usage{
plot_venn_w_genes(diffex_list, mytitle = NULL, selected_phase = NULL)
}
\arguments{
\item{diffex_list}{Parameter for diffex list}

\item{mytitle}{Plot title}

\item{selected_phase}{Parameter for selected phase}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
