% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_52.R
\name{poster_plot_markers}
\alias{poster_plot_markers}
\title{Create a plot visualization}
\usage{
poster_plot_markers(
  seu,
  metavar = "batch",
  num_markers = 5,
  selected_values = NULL,
  return_plotly = FALSE,
  marker_method = "presto",
  seurat_assay = "gene",
  hide_technical = NULL,
  unique_markers = FALSE,
  p_val_cutoff = 1,
  ...
)
}
\arguments{
\item{seu}{Seurat object}

\item{metavar}{Character string (default: "batch")}

\item{num_markers}{Parameter for num markers}

\item{selected_values}{Parameter for selected values}

\item{return_plotly}{Logical flag (default: FALSE)}

\item{marker_method}{Character string (default: "presto")}

\item{seurat_assay}{Character string (default: "gene")}

\item{hide_technical}{Parameter for hide technical}

\item{unique_markers}{Logical flag (default: FALSE)}

\item{p_val_cutoff}{Threshold value for filtering}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
