% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_38.R
\name{prep_for_enrichment}
\alias{prep_for_enrichment}
\title{Perform enrichment analysis}
\usage{
prep_for_enrichment(df, ...)
}
\arguments{
\item{df}{Input data frame or dataset}

\item{...}{Additional arguments passed to other functions}
}
\value{
Enrichment analysis results
}
\description{
Perform enrichment analysis
}
