% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_1.R
\name{prep_seu_branch}
\alias{prep_seu_branch}
\title{Perform prep seu branch operation}
\usage{
prep_seu_branch(debranched_seu, ...)
}
\arguments{
\item{debranched_seu}{Parameter for debranched seu}

\item{...}{Additional arguments passed to other functions}
}
\value{
Modified Seurat object
}
\description{
Perform prep seu branch operation
}
