% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_3.R
\name{prep_unfiltered_seu}
\alias{prep_unfiltered_seu}
\title{Filter data based on specified criteria}
\usage{
prep_unfiltered_seu(
  numbat_rds_file,
  cluster_dictionary,
  large_clone_simplifications,
  filter_expressions = NULL,
  cells_to_remove = NULL,
  extension = ""
)
}
\arguments{
\item{numbat_rds_file}{File path}

\item{cluster_dictionary}{Cluster information}

\item{large_clone_simplifications}{Parameter for large clone simplifications}

\item{filter_expressions}{Parameter for filter expressions}

\item{cells_to_remove}{Cell identifiers or information}

\item{extension}{Character string (default: "")}
}
\value{
Modified Seurat object
}
\description{
Filter data based on specified criteria
}
