% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_52.R
\name{print_table_tally}
\alias{print_table_tally}
\title{Perform print table tally operation}
\usage{
print_table_tally(comparison_id, myvec)
}
\arguments{
\item{comparison_id}{Parameter for comparison id}

\item{myvec}{Parameter for myvec}
}
\value{
Function result
}
\description{
Perform print table tally operation
}
