% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_1.R
\name{pull_branches}
\alias{pull_branches}
\title{Extract or pull specific data elements}
\usage{
pull_branches(branch_dictionary_file = "data/branch_dictionary.csv")
}
\arguments{
\item{branch_dictionary_file}{File path}
}
\value{
Extracted data elements
}
\description{
Extract or pull specific data elements
}
