% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_29.R
\name{pull_cells_matching_expression}
\alias{pull_cells_matching_expression}
\title{Extract or pull specific data elements}
\usage{
pull_cells_matching_expression(myexpression, joint_post)
}
\arguments{
\item{myexpression}{Parameter for myexpression}

\item{joint_post}{Parameter for joint post}
}
\value{
Extracted data elements
}
\description{
Extract or pull specific data elements
}
