% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_46.R
\name{pull_cluster_orders}
\alias{pull_cluster_orders}
\title{Perform clustering analysis}
\usage{
pull_cluster_orders(raw_cluster_file = "data/scna_cluster_order.csv")
}
\arguments{
\item{raw_cluster_file}{File path}
}
\value{
List object
}
\description{
Perform clustering analysis
}
