% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_46.R
\name{pull_cluster_orders2}
\alias{pull_cluster_orders2}
\title{Perform clustering analysis}
\usage{
pull_cluster_orders2(json_file = "data/scna_cluster_order.json")
}
\arguments{
\item{json_file}{File path}
}
\value{
List object
}
\description{
Perform clustering analysis
}
