% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_1.R
\name{pull_clustree_tables}
\alias{pull_clustree_tables}
\title{Extract or pull specific data elements}
\usage{
pull_clustree_tables(
  clustrees,
  divergent_cluster_file = "data/clustree_divergent_clusters.csv"
)
}
\arguments{
\item{clustrees}{Parameter for clustrees}

\item{divergent_cluster_file}{File path}
}
\value{
Extracted data elements
}
\description{
Extract or pull specific data elements
}
