% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_44.R
\name{pull_common_markers}
\alias{pull_common_markers}
\title{Extract or pull specific data elements}
\usage{
pull_common_markers(filtered_seus, gene_lists)
}
\arguments{
\item{filtered_seus}{Parameter for filtered seus}

\item{gene_lists}{Gene names or identifiers}

\item{seu_path}{File path}
}
\value{
Extracted data elements

Extracted data elements
}
\description{
Extract or pull specific data elements
}
