% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_18.R
\name{pull_scna_segments}
\alias{pull_scna_segments}
\title{Extract or pull specific data elements}
\usage{
pull_scna_segments(
  nb_path = "output/numbat_sridhar/SRX11133594_numbat.rds",
  chrom = "1"
)
}
\arguments{
\item{nb_path}{File path}

\item{chrom}{Character string (default: "1")}
}
\value{
Extracted data elements
}
\description{
Extract or pull specific data elements
}
