% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_37.R
\name{pull_stem_cell_markers}
\alias{pull_stem_cell_markers}
\title{Extract or pull specific data elements}
\usage{
pull_stem_cell_markers()
}
\value{
List object
}
\description{
Extract or pull specific data elements
}
