% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_37.R
\name{pull_subtype_genes}
\alias{pull_subtype_genes}
\title{Extract or pull specific data elements}
\usage{
pull_subtype_genes(
  supp_excel = "data/Liu_Radvanyi_2022_supp_data/41467_2021_25792_MOESM6_ESM.xlsx",
  subtype_chr_preference_file = "results/chromosome_distribution_of_subtype_genes.xlsx"
)
}
\arguments{
\item{supp_excel}{Character string (default: "data/Liu_Radvanyi_2022_supp_data/41467_2021_25792_MOESM6_ESM.xlsx")}

\item{subtype_chr_preference_file}{File path}
}
\value{
Extracted data elements
}
\description{
Extract or pull specific data elements
}
