% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_17.R
\name{rb_scna_frequency_in_tcga_by_cancer_type}
\alias{rb_scna_frequency_in_tcga_by_cancer_type}
\title{Perform make fig s16 table s04 operation}
\usage{
rb_scna_frequency_in_tcga_by_cancer_type(
  plot_path = "results/fig_s02.pdf",
  table_path = "results/table_s04.csv",
  ...
)
}
\arguments{
\item{plot_path}{File path}

\item{table_path}{File path}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Perform make fig s16 table s04 operation
}
