% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions_1.R
\name{read_cell_barcodes_from_db}
\alias{read_cell_barcodes_from_db}
\title{Read retained cell barcodes for a Seurat object from the database}
\usage{
read_cell_barcodes_from_db(filepath, sqlite_path = "batch_hashes.sqlite")
}
\arguments{
\item{filepath}{Path to the Seurat RDS file}

\item{sqlite_path}{Path to the SQLite database}
}
\value{
Character vector of cell barcodes, or NULL if not found
}
\description{
Read retained cell barcodes for a Seurat object from the database
}
