% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_43.R
\name{read_giotti_genes}
\alias{read_giotti_genes}
\title{Load or read data from file}
\usage{
read_giotti_genes(cc_file = "data/giotti_cc_genes.tsv")
}
\arguments{
\item{cc_file}{File path}
}
\value{
Loaded data object
}
\description{
Load or read data from file
}
