% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_50.R
\name{read_liu_lu_supp_tables}
\alias{read_liu_lu_supp_tables}
\title{Load or read data from file}
\usage{
read_liu_lu_supp_tables()
}
\value{
Loaded data object
}
\description{
Load or read data from file
}
