% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_10.R
\name{read_mps}
\alias{read_mps}
\title{Load or read data from file}
\usage{
read_mps(
  mp_file =
    "/dataVolume/storage/Homo_sapiens/3ca/ITH_hallmarks/MPs_distribution/MP_list.RDS"
)
}
\arguments{
\item{mp_file}{File path}
}
\value{
Loaded data object
}
\description{
Load or read data from file
}
