% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_43.R
\name{read_zinovyev_genes}
\alias{read_zinovyev_genes}
\title{Load or read data from file}
\usage{
read_zinovyev_genes(zinovyev_file = "data/zinovyev_cc_genes.tsv")
}
\arguments{
\item{zinovyev_file}{File path}
}
\value{
Loaded data object
}
\description{
Load or read data from file
}
