% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_42.R
\name{reference_plae_celltypes}
\alias{reference_plae_celltypes}
\title{Perform reference plae celltypes operation}
\usage{
reference_plae_celltypes(seu_path, mycluster = "SCT_snn_res.0.4", mygenes)
}
\arguments{
\item{seu_path}{File path}

\item{mycluster}{Cluster information}

\item{mygenes}{Gene names or identifiers}
}
\value{
Function result
}
\description{
Perform reference plae celltypes operation
}
