% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_3.R
\name{regress_filtered_seu}
\alias{regress_filtered_seu}
\title{Filter data based on specified criteria}
\usage{
regress_filtered_seu(filtered_seu_path)
}
\arguments{
\item{filtered_seu_path}{File path}
}
\value{
Modified Seurat object
}
\description{
Filter data based on specified criteria
}
