% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_33.R
\name{remove_non_tumor_cells}
\alias{remove_non_tumor_cells}
\title{Perform remove non tumor cells operation}
\usage{
remove_non_tumor_cells(seu)
}
\arguments{
\item{seu}{Seurat object}
}
\value{
Function result
}
\description{
Perform remove non tumor cells operation
}
