% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_17.R
\name{reorder_within}
\alias{reorder_within}
\title{Perform reorder within operation}
\usage{
reorder_within(x, by, within, fun = mean, sep = "___", ...)
}
\arguments{
\item{x}{Parameter for x}

\item{by}{Parameter for by}

\item{within}{Parameter for within}

\item{fun}{Parameter for fun}

\item{sep}{Character string (default: "___")}

\item{...}{Additional arguments passed to other functions}
}
\value{
Function result
}
\description{
Perform reorder within operation
}
