% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_37.R
\name{rescale_and_clean_plots}
\alias{rescale_and_clean_plots}
\title{Create a plot visualization}
\usage{
rescale_and_clean_plots(groblist)
}
\arguments{
\item{groblist}{Parameter for groblist}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
