% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_10.R
\name{retrieve_cell_stats}
\alias{retrieve_cell_stats}
\title{Perform retrieve cell stats operation}
\usage{
retrieve_cell_stats(seu_path, sqlite_path = "batch_hashes.sqlite")
}
\arguments{
\item{seu_path}{File path}

\item{sqlite_path}{Path to Seurat metadata SQLite database}
}
\value{
Function result
}
\description{
Perform retrieve cell stats operation
}
