% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions_1.R
\name{retrieve_current_param}
\alias{retrieve_current_param}
\title{Perform retrieve current param operation}
\usage{
retrieve_current_param(current_params, myparam)
}
\arguments{
\item{current_params}{Parameter for current params}

\item{myparam}{Parameter for myparam}
}
\value{
Named list of param values by sample ID
}
\description{
Perform retrieve current param operation
}
