% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_13.R
\name{retrieve_genes_in_cis}
\alias{retrieve_genes_in_cis}
\title{Perform retrieve genes in cis operation}
\usage{
retrieve_genes_in_cis(nb_path, tumor_id = "asdf")
}
\arguments{
\item{nb_path}{File path}

\item{tumor_id}{Character string (default: "asdf")}
}
\value{
Function result
}
\description{
Perform retrieve genes in cis operation
}
