% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_6.R
\name{retrieve_numbat_seurat}
\alias{retrieve_numbat_seurat}
\title{Title}
\usage{
retrieve_numbat_seurat(numbat_rds_file, cluster_dictionary)
}
\arguments{
\item{cluster_dictionary}{}
}
\description{
Title
}
