% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_14.R
\name{retrieve_segmentation}
\alias{retrieve_segmentation}
\title{Perform retrieve segmentation operation}
\usage{
retrieve_segmentation(rds_file)
}
\arguments{
\item{rds_file}{File path}
}
\value{
Function result
}
\description{
Perform retrieve segmentation operation
}
