% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions_1.R
\name{retrieve_snakemake_params}
\alias{retrieve_snakemake_params}
\title{Perform retrieve snakemake params operation}
\usage{
retrieve_snakemake_params(numbat_rds_file)
}
\arguments{
\item{numbat_rds_file}{File path}
}
\value{
List object
}
\description{
Perform retrieve snakemake params operation
}
