% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_45.R
\name{run_speckle}
\alias{run_speckle}
\title{Perform run speckle operation}
\usage{
run_speckle(
  seu_path = "output/seurat/SRX11133594_filtered_seu.rds",
  group.by = "SCT_snn_res.0.6"
)
}
\arguments{
\item{seu_path}{File path}

\item{group.by}{Character string (default: "SCT_snn_res.0.6")}
}
\value{
ggplot2 plot object
}
\description{
Perform run speckle operation
}
