% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_45.R
\name{run_speckle_for_set}
\alias{run_speckle_for_set}
\title{Perform run speckle for set operation}
\usage{
run_speckle_for_set(seu_paths, group.by = "SCT_snn_res.0.6")
}
\arguments{
\item{seu_paths}{File path}

\item{group.by}{Character string (default: "SCT_snn_res.0.6")}
}
\value{
Function result
}
\description{
Perform run speckle for set operation
}
