% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_50.R
\name{run_velocity}
\alias{run_velocity}
\title{Perform run velocity operation}
\usage{
run_velocity(debranched_seu)
}
\arguments{
\item{debranched_seu}{Parameter for debranched seu}
}
\value{
Function result
}
\description{
Perform run velocity operation
}
