% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_48.R
\name{save_cc_space_plot_from_path}
\alias{save_cc_space_plot_from_path}
\title{Create a plot visualization}
\usage{
save_cc_space_plot_from_path(
  seu_path,
  clone_simplifications,
  label = "_clone_tree",
  ...
)
}
\arguments{
\item{seu_path}{File path}

\item{clone_simplifications}{Parameter for clone simplifications}

\item{label}{Character string (default: "_clone_tree")}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
