% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_47.R
\name{save_clone_tree_from_path}
\alias{save_clone_tree_from_path}
\title{Save or write data to file}
\usage{
save_clone_tree_from_path(
  seu_path,
  nb_paths,
  clone_simplifications,
  label = "_clone_tree",
  ...
)
}
\arguments{
\item{seu_path}{File path}

\item{nb_paths}{File path}

\item{clone_simplifications}{Parameter for clone simplifications}

\item{label}{Character string (default: "_clone_tree")}

\item{...}{Additional arguments passed to other functions}
}
\value{
path to a ggplot pdf file
}
\description{
Save or write data to file
}
