% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_17.R
\name{scale_x_reordered}
\alias{scale_x_reordered}
\title{Perform scale x reordered operation}
\usage{
scale_x_reordered(..., sep = "___")
}
\arguments{
\item{...}{Additional arguments passed to other functions}

\item{sep}{Character string (default: "___")}
}
\value{
Function result
}
\description{
Perform scale x reordered operation
}
