% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_37.R
\name{score_and_heatmap_seu}
\alias{score_and_heatmap_seu}
\title{Calculate scores for the given data}
\usage{
score_and_heatmap_seu(
  seu_path,
  gene_lists,
  group.by = "SCT_snn_res.0.4",
  leiden_cluster_file = NULL
)
}
\arguments{
\item{seu_path}{File path}

\item{gene_lists}{Gene names or identifiers}

\item{group.by}{Character string (default: "SCT_snn_res.0.4")}

\item{leiden_cluster_file}{File path}
}
\value{
ggplot2 plot object
}
\description{
Calculate scores for the given data
}
