% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_36.R
\name{score_and_vlnplot_seu}
\alias{score_and_vlnplot_seu}
\title{Create a plot visualization}
\usage{
score_and_vlnplot_seu(
  seu_path,
  nb_path,
  clone_simplifications,
  gene_lists,
  ...
)
}
\arguments{
\item{seu_path}{File path}

\item{nb_path}{File path}

\item{clone_simplifications}{Parameter for clone simplifications}

\item{gene_lists}{Gene names or identifiers}

\item{...}{Additional arguments passed to other functions}

\item{score.by}{Character string (default: "subtype2")}

\item{group.by}{Character string (default: "scna")}

\item{seu}{Seurat object}

\item{y_lim}{Parameter for y lim}

\item{step}{Parameter for step}
}
\value{
ggplot2 plot object

Function result
}
\description{
Create a plot visualization
}
