% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_41.R
\name{score_binary_celltype_clusters}
\alias{score_binary_celltype_clusters}
\title{Calculate scores for the given data}
\usage{
score_binary_celltype_clusters(seu, celltype_markers)
}
\arguments{
\item{seu}{Seurat object}

\item{celltype_markers}{Cell identifiers or information}

\item{upper_quartile}{Parameter for upper quartile}

\item{cluster_name}{Cluster information}
}
\value{
Modified Seurat object

Function result
}
\description{
Calculate scores for the given data
}
