% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_41.R
\name{score_binary_celltype_markers}
\alias{score_binary_celltype_markers}
\title{Calculate scores for the given data}
\usage{
score_binary_celltype_markers(seu, celltype_markers)
}
\arguments{
\item{seu}{Seurat object}

\item{celltype_markers}{Cell identifiers or information}
}
\value{
Modified Seurat object
}
\description{
Calculate scores for the given data
}
