% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R, R/plot_functions_42.R
\name{score_chrom_instability}
\alias{score_chrom_instability}
\title{Score chromosomal instability}
\usage{
score_chrom_instability(seu_path)

score_chrom_instability(seu_path)
}
\arguments{
\item{seu_path}{File path}

\item{seu}{Seurat object}
}
\value{
Numeric vector of instability scores

ggplot2 plot object
}
\description{
Score chromosomal instability

Calculate scores for the given data
}
