% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_47.R
\name{score_clusters_up_down}
\alias{score_clusters_up_down}
\title{Calculate scores for the given data}
\usage{
score_clusters_up_down(
  clone_distribution_plots,
  csv_file = "results/cluster_up_down_scorecard.csv"
)
}
\arguments{
\item{clone_distribution_plots}{Parameter for clone distribution plots}

\item{csv_file}{File path}

\item{interesting_samples}{Parameter for interesting samples}

\item{df}{Input data frame or dataset}
}
\value{
Numeric scores or scored data

Function result
}
\description{
Calculate scores for the given data
}
