% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_29.R
\name{score_filtration}
\alias{score_filtration}
\title{Calculate scores for the given data}
\usage{
score_filtration(
  numbat_rds_file,
  cluster_dictionary,
  filter_expressions = NULL
)
}
\arguments{
\item{numbat_rds_file}{File path}

\item{cluster_dictionary}{Cluster information}

\item{filter_expressions}{Parameter for filter expressions}
}
\value{
List object
}
\description{
Calculate scores for the given data
}
