% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_41.R
\name{score_samples_for_celltype_enrichment}
\alias{score_samples_for_celltype_enrichment}
\title{Calculate scores for the given data}
\usage{
score_samples_for_celltype_enrichment(
  unfiltered_seu_path,
  filtered_seu_path,
  celltype_markers
)
}
\arguments{
\item{unfiltered_seu_path}{File path}

\item{filtered_seu_path}{File path}

\item{celltype_markers}{Cell identifiers or information}
}
\value{
ggplot2 plot object
}
\description{
Calculate scores for the given data
}
