% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_39.R
\name{score_samples_for_rod_enrichment}
\alias{score_samples_for_rod_enrichment}
\title{Calculate scores for the given data}
\usage{
score_samples_for_rod_enrichment(numbat_rds_file)
}
\arguments{
\item{numbat_rds_file}{File path}
}
\value{
List object
}
\description{
Calculate scores for the given data
}
