% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_42.R
\name{score_stachelek}
\alias{score_stachelek}
\title{Calculate scores for the given data}
\usage{
score_stachelek(seu_path, stachelek_scores_table)
}
\arguments{
\item{seu_path}{File path}

\item{stachelek_scores_table}{Parameter for stachelek scores table}
}
\value{
ggplot2 plot object
}
\description{
Calculate scores for the given data
}
