% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_39.R
\name{score_whole_pseudobulks}
\alias{score_whole_pseudobulks}
\title{Calculate scores for the given data}
\usage{
score_whole_pseudobulks(numbat_rds_file, subtype_markers)
}
\arguments{
\item{numbat_rds_file}{File path}

\item{subtype_markers}{Parameter for subtype markers}
}
\value{
List object
}
\description{
Calculate scores for the given data
}
