% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_49.R
\name{select_genes_from_arbitrary_diffex}
\alias{select_genes_from_arbitrary_diffex}
\title{Perform differential expression analysis}
\usage{
select_genes_from_arbitrary_diffex(oncoprint_input_by_scna_unfiltered)
}
\arguments{
\item{oncoprint_input_by_scna_unfiltered}{Parameter for oncoprint input by scna unfiltered}

\item{df}{Input data frame or dataset}
}
\value{
Differential expression results

Function result
}
\description{
Perform differential expression analysis
}
