% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_12.R
\name{select_genes_to_plot}
\alias{select_genes_to_plot}
\title{Create a plot visualization}
\usage{
select_genes_to_plot(
  seu,
  direction,
  mygenes,
  max_genes = 15,
  scna_of_interest = "02_p",
  in_scna = TRUE,
  only_scna = FALSE
)
}
\arguments{
\item{seu}{Seurat object}

\item{direction}{Parameter for direction}

\item{mygenes}{Gene names or identifiers}

\item{max_genes}{Gene names or identifiers}

\item{scna_of_interest}{Character string (default: "02_p")}

\item{in_scna}{Logical flag (default: TRUE)}

\item{only_scna}{Logical flag (default: FALSE)}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
