% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_50.R
\name{set_final_seus}
\alias{set_final_seus}
\title{Perform set final seus operation}
\usage{
set_final_seus(interesting_samples)
}
\arguments{
\item{interesting_samples}{Parameter for interesting samples}
}
\value{
Function result
}
\description{
Perform set final seus operation
}
