% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_53.R
\name{seu_complex_heatmap}
\alias{seu_complex_heatmap}
\title{Perform seu complex heatmap operation}
\usage{
seu_complex_heatmap(
  seu,
  features = NULL,
  group.by = "ident",
  cells = NULL,
  layer = "scale.data",
  assay = NULL,
  group.bar.height = 0.01,
  column_split = NULL,
  col_arrangement = "ward.D2",
  mm_col_dend = 30,
  embedding = "pca",
  ...
)
}
\arguments{
\item{seu}{Seurat object}

\item{features}{Parameter for features}

\item{group.by}{Character string (default: "ident")}

\item{cells}{Cell identifiers or information}

\item{layer}{Character string (default: "scale.data")}

\item{assay}{Parameter for assay}

\item{group.bar.height}{Parameter for group.bar.height}

\item{column_split}{Color specification}

\item{col_arrangement}{Color specification}

\item{mm_col_dend}{Color specification}

\item{embedding}{Character string (default: "pca")}

\item{...}{Additional arguments passed to other functions}
}
\value{
Function result
}
\description{
Perform seu complex heatmap operation
}
