% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_53.R
\name{seu_gene_heatmap}
\alias{seu_gene_heatmap}
\title{Perform seu gene heatmap operation}
\usage{
seu_gene_heatmap(
  seu,
  marker_col = "clusters",
  group.by = "ident",
  cells = NULL,
  layer = "scale.data",
  assay = NULL,
  group.bar.height = 0.01,
  column_split = NULL,
  row_split = NULL,
  col_arrangement = "ward.D2",
  mm_col_dend = 30,
  embedding = "pca",
  factor_cols = NULL,
  summarize = FALSE,
  hide_legends = NULL,
  heatmap_features = NULL,
  seurat_assay = "SCT",
  features = NULL,
  cluster_rows = FALSE,
  ...
)
}
\arguments{
\item{seu}{Seurat object}

\item{marker_col}{Color specification}

\item{group.by}{Character string (default: "ident")}

\item{cells}{Cell identifiers or information}

\item{layer}{Character string (default: "scale.data")}

\item{assay}{Parameter for assay}

\item{group.bar.height}{Parameter for group.bar.height}

\item{column_split}{Color specification}

\item{row_split}{Parameter for row split}

\item{col_arrangement}{Color specification}

\item{mm_col_dend}{Color specification}

\item{embedding}{Character string (default: "pca")}

\item{factor_cols}{Color specification}

\item{summarize}{Logical flag (default: FALSE)}

\item{hide_legends}{Parameter for hide legends}

\item{heatmap_features}{Parameter for heatmap features}

\item{seurat_assay}{Character string (default: "SCT")}

\item{features}{Parameter for features}

\item{cluster_rows}{Cluster information}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot2 plot object
}
\description{
Perform seu gene heatmap operation
}
