% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_1.R
\name{split_by_cnv}
\alias{split_by_cnv}
\title{Perform split by cnv operation}
\usage{
split_by_cnv(returned_meta, myseg = "2a")
}
\arguments{
\item{returned_meta}{Parameter for returned meta}

\item{myseg}{Character string (default: "2a")}
}
\value{
Function result
}
\description{
Perform split by cnv operation
}
