% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_41.R
\name{split_label_line}
\alias{split_label_line}
\title{Perform split label line operation}
\usage{
split_label_line(label, n_comma_values = 2)
}
\arguments{
\item{label}{Parameter for label}

\item{n_comma_values}{Parameter for n comma values}
}
\value{
Function result
}
\description{
Perform split label line operation
}
