% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_1.R
\name{split_seu_by_branch}
\alias{split_seu_by_branch}
\title{Perform split seu by branch operation}
\usage{
split_seu_by_branch(seu_path, branches, ...)
}
\arguments{
\item{seu_path}{File path}

\item{branches}{Parameter for branches}

\item{...}{Additional arguments passed to other functions}
}
\value{
Modified Seurat object
}
\description{
Perform split seu by branch operation
}
