% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_20.R
\name{split_vector_into_overlapping_chunks}
\alias{split_vector_into_overlapping_chunks}
\title{Perform split vector into overlapping chunks operation}
\usage{
split_vector_into_overlapping_chunks(v, n)
}
\arguments{
\item{v}{Parameter for v}

\item{n}{Parameter for n}
}
\value{
Function result
}
\description{
Perform split vector into overlapping chunks operation
}
