% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{subset_to_16q}
\alias{subset_to_16q}
\title{Select 1q clone Seurat objects and save}
\usage{
subset_to_16q(seu, file_id = NULL, slug = "", ...)
}
\arguments{
\item{seu}{A Seurat object.}

\item{slug}{Optional slug appended to output filenames.}

\item{...}{Additional arguments passed to downstream plotting functions.}
}
\value{
Path to the generated PDF clone distribution figure.
}
\description{
Wrapper that subsets the provided Seurat object to 1q clones, saves the
resulting Seurat RDS and creates a clone distribution figure.
}
