% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_51.R
\name{table_and_plot_enrichment}
\alias{table_and_plot_enrichment}
\title{Create a plot visualization}
\usage{
table_and_plot_enrichment(ident.1, ident.2, scna_of_interest, seu, sample_id)
}
\arguments{
\item{ident.1}{Cell identities or groups}

\item{ident.2}{Cell identities or groups}

\item{scna_of_interest}{Parameter for scna of interest}

\item{seu}{Seurat object}

\item{sample_id}{Parameter for sample id}
}
\value{
List object
}
\description{
Create a plot visualization
}
