% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_22.R
\name{table_cluster_markers}
\alias{table_cluster_markers}
\title{Perform clustering analysis}
\usage{
table_cluster_markers(seu, assay = "SCT")
}
\arguments{
\item{seu}{Seurat object}

\item{assay}{Character string (default: "SCT")}
}
\value{
Function result
}
\description{
Perform clustering analysis
}
