% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_22.R
\name{table_distribution_of_clones_across_clusters}
\alias{table_distribution_of_clones_across_clusters}
\title{Perform clustering analysis}
\usage{
table_distribution_of_clones_across_clusters(
  seu,
  seu_name,
  clusters = "SCT_snn_res.0.4"
)
}
\arguments{
\item{seu}{Seurat object}

\item{seu_name}{Parameter for seu name}

\item{clusters}{Cluster information}
}
\value{
List object
}
\description{
Perform clustering analysis
}
