% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_14.R
\name{tabulate_clone_percent}
\alias{tabulate_clone_percent}
\title{Perform tabulate clone percent operation}
\usage{
tabulate_clone_percent(seu_path = "output/seurat/SRX11133594_filtered_seu.rds")
}
\arguments{
\item{seu_path}{File path}
}
\value{
Function result
}
\description{
Perform tabulate clone percent operation
}
