% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbat_functions_10.R
\name{tabulate_clone_percent_integrated}
\alias{tabulate_clone_percent_integrated}
\title{Perform tabulate clone percent integrated operation}
\usage{
tabulate_clone_percent_integrated(
  seu_path = "output/seurat/integrated_1q/integrated_seu_1q_complete.rds"
)
}
\arguments{
\item{seu_path}{File path}
}
\value{
Function result
}
\description{
Perform tabulate clone percent integrated operation
}
