% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_32.R
\name{tabulate_diffex_clones}
\alias{tabulate_diffex_clones}
\title{Perform differential expression analysis}
\usage{
tabulate_diffex_clones(
  cluster_diffex_clones,
  cluster_xlsx = "results/diffex_bw_clones_per_cluster_large.xlsx",
  cluster_by_chr_xlsx = "results/diffex_bw_clones_per_cluster_large_by_chr.xlsx",
  total_diffex_clones,
  total_xlsx = "results/straight_diffex_bw_clones_large.xlsx",
  total_by_chr_xlsx = "results/straight_diffex_bw_clones_large_by_chr.xlsx"
)
}
\arguments{
\item{cluster_diffex_clones}{Cluster information}

\item{cluster_xlsx}{Cluster information}

\item{cluster_by_chr_xlsx}{Cluster information}

\item{total_diffex_clones}{Parameter for total diffex clones}

\item{total_xlsx}{Character string (default: "results/straight_diffex_bw_clones_large.xlsx")}

\item{total_by_chr_xlsx}{Character string (default: "results/straight_diffex_bw_clones_large_by_chr.xlsx")}

\item{diffex}{Parameter for diffex}
}
\value{
Differential expression results

Differential expression results
}
\description{
Perform differential expression analysis
}
