% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_10.R
\name{tally_kooi_candidates}
\alias{tally_kooi_candidates}
\title{Perform tally kooi candidates operation}
\usage{
tally_kooi_candidates(
  cis_diffex_clones = "results/diffex_bw_clones_large_in_segment_by_chr.xlsx",
  trans_diffex_clones = "results/diffex_bw_clones_trans_by_chr.xlsx"
)
}
\arguments{
\item{cis_diffex_clones}{Character string (default: "results/diffex_bw_clones_large_in_segment_by_chr.xlsx")}

\item{trans_diffex_clones}{Character string (default: "results/diffex_bw_clones_trans_by_chr.xlsx")}
}
\value{
List object
}
\description{
Perform tally kooi candidates operation
}
