% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions_10.R
\name{tally_num_diffex}
\alias{tally_num_diffex}
\title{Perform differential expression analysis}
\usage{
tally_num_diffex(oncoprint_input_by_scna_unfiltered)
}
\arguments{
\item{oncoprint_input_by_scna_unfiltered}{Parameter for oncoprint input by scna unfiltered}
}
\value{
List object
}
\description{
Perform differential expression analysis
}
