% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_45.R
\name{tidy_eval_arrange}
\alias{tidy_eval_arrange}
\title{Perform tidy eval arrange operation}
\usage{
tidy_eval_arrange(.data, ...)
}
\arguments{
\item{.data}{Parameter for .data}

\item{...}{Additional arguments passed to other functions}
}
\value{
Function result
}
\description{
Perform tidy eval arrange operation
}
