% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_32.R
\name{volcano_plot_clone_clusters}
\alias{volcano_plot_clone_clusters}
\title{Generate a volcano plot for differential expression data}
\usage{
volcano_plot_clone_clusters(clone_diffex, sample_id)
}
\arguments{
\item{clone_diffex}{Parameter for clone diffex}

\item{sample_id}{Parameter for sample id}
}
\value{
ggplot2 plot object
}
\description{
Generate a volcano plot for differential expression data
}
