% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_26.R
\name{window}
\alias{window}
\title{Perform window operation}
\usage{
window(x, size)
}
\arguments{
\item{x}{Parameter for x}

\item{size}{Size parameter}
}
\value{
Function result
}
\description{
Perform window operation
}
