% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_44.R
\name{wrap_scna_labels}
\alias{wrap_scna_labels}
\title{Perform wrap scna labels operation}
\usage{
wrap_scna_labels(scna_labels)
}
\arguments{
\item{scna_labels}{Parameter for scna labels}
}
\value{
Function result
}
\description{
Perform wrap scna labels operation
}
