% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{make_numbat_heatmaps}
\alias{make_numbat_heatmaps}
\title{Make numbat heatmaps from Seurat and numbat objects}
\usage{
make_numbat_heatmaps(
  seu,
  numbat_rds_files,
  p_min = 0.9,
  line_width = 0.1,
  extension = "",
  midline_threshold = 0.4,
  show_segment_names_on_x = FALSE,
  numbat_rds_filtered_files = NULL
)
}
\arguments{
\item{seu}{Seurat object}

\item{numbat_rds_files}{Numbat RDS file paths}

\item{p_min}{Minimum p value}

\item{line_width}{Line width for plots}

\item{extension}{File extension}

\item{midline_threshold}{Threshold for midline filtering}

\item{show_segment_names_on_x}{Show segment names on x-axis}

\item{numbat_rds_filtered_files}{Optional filtered numbat files}
}
\value{
Character vector with plot paths
}
\description{
Make numbat heatmaps from Seurat and numbat objects
}
