% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_48.R
\name{plot_clone_tree}
\alias{plot_clone_tree}
\title{Create a plot visualization}
\usage{
plot_clone_tree(
  clone_df,
  tumor_id,
  nb_path,
  clone_simplifications = NULL,
  sample_id = NULL,
  show_distance = FALSE,
  ...
)
}
\arguments{
\item{tumor_id}{Parameter for tumor id}

\item{nb_path}{File path}

\item{clone_simplifications}{Parameter for clone simplifications}

\item{sample_id}{Parameter for sample id}

\item{...}{Additional arguments passed to other functions}

\item{seu}{Seurat object}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
