% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R, R/plot_functions_1.R
\name{plot_numbat_w_phylo}
\alias{plot_numbat_w_phylo}
\title{Plot numbat heatmap with phylogeny}
\usage{
plot_numbat_w_phylo(
  nb,
  myseu,
  myannot,
  mytitle,
  show_segment_names_on_x = FALSE,
  ...
)

plot_numbat_w_phylo(
  nb,
  myseu,
  myannot,
  mytitle,
  show_segment_names_on_x = FALSE,
  ...
)
}
\arguments{
\item{nb}{Numbat object}

\item{myseu}{Seurat object}

\item{myannot}{Parameter for myannot}

\item{mytitle}{Plot title}

\item{show_segment_names_on_x}{Logical; if TRUE, keep segment labels on applicable genomic regions.}

\item{...}{Additional arguments passed to other functions}
}
\value{
ggplot object

ggplot2 plot object
}
\description{
Plot numbat heatmap with phylogeny

Create a numbat-related plot visualization
}
