% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_variability_at_SCNA}
\alias{plot_variability_at_SCNA}
\title{Plot variability at SCNA}
\usage{
plot_variability_at_SCNA(phylo_plot_output, chrom = "1", p_min = 0.9)
}
\arguments{
\item{phylo_plot_output}{Data from phylo plot}

\item{chrom}{Chromosome}

\item{p_min}{Minimum p value}
}
\value{
ggplot object
}
\description{
Plot variability at SCNA
}
