% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_functions.R, R/plot_functions_13.R
\name{plot_venn_w_genes}
\alias{plot_venn_w_genes}
\title{Create a Venn diagram of recurrent diffex genes}
\usage{
plot_venn_w_genes(diffex_list, mytitle = NULL, selected_phase = NULL)

plot_venn_w_genes(diffex_list, mytitle = NULL, selected_phase = NULL)
}
\arguments{
\item{diffex_list}{Parameter for diffex list}

\item{mytitle}{Plot title}

\item{selected_phase}{Parameter for selected phase}
}
\value{
List with ggplot2 Venn plot and region data

ggplot2 plot object
}
\description{
Create a Venn diagram of recurrent diffex genes

Create a plot visualization
}
