% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_15.R
\name{sample_specific_analyses_of_tumors_with_scna_subclones_after_integration}
\alias{sample_specific_analyses_of_tumors_with_scna_subclones_after_integration}
\title{Create a plot visualization}
\usage{
sample_specific_analyses_of_tumors_with_scna_subclones_after_integration(
  seu_list,
  cluster_orders,
  plot_path = "results/fig_s04.pdf"
)
}
\arguments{
\item{seu_list}{Parameter for seu list}

\item{cluster_orders}{Cluster information}

\item{plot_path}{File path}
}
\value{
ggplot2 plot object
}
\description{
Create a plot visualization
}
