% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{score_chrom_instability}
\alias{score_chrom_instability}
\title{Score chromosomal instability}
\usage{
score_chrom_instability(seu)
}
\arguments{
\item{seu}{Seurat object}
}
\value{
Numeric vector of instability scores
}
\description{
Score chromosomal instability
}
