% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_42.R
\name{score_chrom_instability}
\alias{score_chrom_instability}
\title{Score chromosomal instability using CIN gene signatures}
\usage{
score_chrom_instability(seu_path)
}
\arguments{
\item{seu_path}{File path to Seurat RDS, or list containing a path}
}
\value{
File path of output PDF
}
\description{
Score chromosomal instability using CIN gene signatures
}
